Address PR review comments. Adjust part d contract history so only 20% of benes change their current plan each Jan.
