+++ src/main/java/org/mitre/synthea/export/BB2RIFStructure.java
@@ -1,0 +1,1859 @@
+package org.mitre.synthea.export;
+
+/**
+ * A collection of enums and related that describe the structure of RIF files. Note that order 
+ * is significant, columns will be written in the order specified.
+ */
+public class BB2RIFStructure {
+
+  enum BeneficiaryFields {
+    DML_IND,
+    BENE_ID,
+    STATE_CODE,
+    BENE_COUNTY_CD,
+    BENE_ZIP_CD,
+    BENE_BIRTH_DT,
+    BENE_SEX_IDENT_CD,
+    BENE_RACE_CD,
+    BENE_ENTLMT_RSN_ORIG,
+    BENE_ENTLMT_RSN_CURR,
+    BENE_ESRD_IND,
+    BENE_MDCR_STATUS_CD,
+    BENE_PTA_TRMNTN_CD,
+    BENE_PTB_TRMNTN_CD,
+    // BENE_PTD_TRMNTN_CD, // The spreadsheet has a gap for this column, examples do not include it
+    BENE_CRNT_HIC_NUM,
+    BENE_SRNM_NAME,
+    BENE_GVN_NAME,
+    BENE_MDL_NAME,
+    MBI_NUM,
+    DEATH_DT,
+    RFRNC_YR,
+    A_MO_CNT,
+    B_MO_CNT,
+    BUYIN_MO_CNT,
+    HMO_MO_CNT,
+    RDS_MO_CNT,
+    ENRL_SRC,
+    SAMPLE_GROUP,
+    EFIVEPCT,
+    CRNT_BIC,
+    AGE,
+    COVSTART,
+    DUAL_MO_CNT,
+    FIPS_STATE_CNTY_JAN_CD,
+    FIPS_STATE_CNTY_FEB_CD,
+    FIPS_STATE_CNTY_MAR_CD,
+    FIPS_STATE_CNTY_APR_CD,
+    FIPS_STATE_CNTY_MAY_CD,
+    FIPS_STATE_CNTY_JUN_CD,
+    FIPS_STATE_CNTY_JUL_CD,
+    FIPS_STATE_CNTY_AUG_CD,
+    FIPS_STATE_CNTY_SEPT_CD,
+    FIPS_STATE_CNTY_OCT_CD,
+    FIPS_STATE_CNTY_NOV_CD,
+    FIPS_STATE_CNTY_DEC_CD,
+    V_DOD_SW,
+    RTI_RACE_CD,
+    MDCR_STUS_JAN_CD,
+    MDCR_STUS_FEB_CD,
+    MDCR_STUS_MAR_CD,
+    MDCR_STUS_APR_CD,
+    MDCR_STUS_MAY_CD,
+    MDCR_STUS_JUN_CD,
+    MDCR_STUS_JUL_CD,
+    MDCR_STUS_AUG_CD,
+    MDCR_STUS_SEPT_CD,
+    MDCR_STUS_OCT_CD,
+    MDCR_STUS_NOV_CD,
+    MDCR_STUS_DEC_CD,
+    PLAN_CVRG_MO_CNT,
+    MDCR_ENTLMT_BUYIN_1_IND,
+    MDCR_ENTLMT_BUYIN_2_IND,
+    MDCR_ENTLMT_BUYIN_3_IND,
+    MDCR_ENTLMT_BUYIN_4_IND,
+    MDCR_ENTLMT_BUYIN_5_IND,
+    MDCR_ENTLMT_BUYIN_6_IND,
+    MDCR_ENTLMT_BUYIN_7_IND,
+    MDCR_ENTLMT_BUYIN_8_IND,
+    MDCR_ENTLMT_BUYIN_9_IND,
+    MDCR_ENTLMT_BUYIN_10_IND,
+    MDCR_ENTLMT_BUYIN_11_IND,
+    MDCR_ENTLMT_BUYIN_12_IND,
+    HMO_1_IND,
+    HMO_2_IND,
+    HMO_3_IND,
+    HMO_4_IND,
+    HMO_5_IND,
+    HMO_6_IND,
+    HMO_7_IND,
+    HMO_8_IND,
+    HMO_9_IND,
+    HMO_10_IND,
+    HMO_11_IND,
+    HMO_12_IND,
+    PTC_CNTRCT_JAN_ID,
+    PTC_CNTRCT_FEB_ID,
+    PTC_CNTRCT_MAR_ID,
+    PTC_CNTRCT_APR_ID,
+    PTC_CNTRCT_MAY_ID,
+    PTC_CNTRCT_JUN_ID,
+    PTC_CNTRCT_JUL_ID,
+    PTC_CNTRCT_AUG_ID,
+    PTC_CNTRCT_SEPT_ID,
+    PTC_CNTRCT_OCT_ID,
+    PTC_CNTRCT_NOV_ID,
+    PTC_CNTRCT_DEC_ID,
+    PTC_PBP_JAN_ID,
+    PTC_PBP_FEB_ID,
+    PTC_PBP_MAR_ID,
+    PTC_PBP_APR_ID,
+    PTC_PBP_MAY_ID,
+    PTC_PBP_JUN_ID,
+    PTC_PBP_JUL_ID,
+    PTC_PBP_AUG_ID,
+    PTC_PBP_SEPT_ID,
+    PTC_PBP_OCT_ID,
+    PTC_PBP_NOV_ID,
+    PTC_PBP_DEC_ID,
+    PTC_PLAN_TYPE_JAN_CD,
+    PTC_PLAN_TYPE_FEB_CD,
+    PTC_PLAN_TYPE_MAR_CD,
+    PTC_PLAN_TYPE_APR_CD,
+    PTC_PLAN_TYPE_MAY_CD,
+    PTC_PLAN_TYPE_JUN_CD,
+    PTC_PLAN_TYPE_JUL_CD,
+    PTC_PLAN_TYPE_AUG_CD,
+    PTC_PLAN_TYPE_SEPT_CD,
+    PTC_PLAN_TYPE_OCT_CD,
+    PTC_PLAN_TYPE_NOV_CD,
+    PTC_PLAN_TYPE_DEC_CD,
+    PTD_CNTRCT_JAN_ID,
+    PTD_CNTRCT_FEB_ID,
+    PTD_CNTRCT_MAR_ID,
+    PTD_CNTRCT_APR_ID,
+    PTD_CNTRCT_MAY_ID,
+    PTD_CNTRCT_JUN_ID,
+    PTD_CNTRCT_JUL_ID,
+    PTD_CNTRCT_AUG_ID,
+    PTD_CNTRCT_SEPT_ID,
+    PTD_CNTRCT_OCT_ID,
+    PTD_CNTRCT_NOV_ID,
+    PTD_CNTRCT_DEC_ID,
+    PTD_PBP_JAN_ID,
+    PTD_PBP_FEB_ID,
+    PTD_PBP_MAR_ID,
+    PTD_PBP_APR_ID,
+    PTD_PBP_MAY_ID,
+    PTD_PBP_JUN_ID,
+    PTD_PBP_JUL_ID,
+    PTD_PBP_AUG_ID,
+    PTD_PBP_SEPT_ID,
+    PTD_PBP_OCT_ID,
+    PTD_PBP_NOV_ID,
+    PTD_PBP_DEC_ID,
+    PTD_SGMT_JAN_ID,
+    PTD_SGMT_FEB_ID,
+    PTD_SGMT_MAR_ID,
+    PTD_SGMT_APR_ID,
+    PTD_SGMT_MAY_ID,
+    PTD_SGMT_JUN_ID,
+    PTD_SGMT_JUL_ID,
+    PTD_SGMT_AUG_ID,
+    PTD_SGMT_SEPT_ID,
+    PTD_SGMT_OCT_ID,
+    PTD_SGMT_NOV_ID,
+    PTD_SGMT_DEC_ID,
+    RDS_JAN_IND,
+    RDS_FEB_IND,
+    RDS_MAR_IND,
+    RDS_APR_IND,
+    RDS_MAY_IND,
+    RDS_JUN_IND,
+    RDS_JUL_IND,
+    RDS_AUG_IND,
+    RDS_SEPT_IND,
+    RDS_OCT_IND,
+    RDS_NOV_IND,
+    RDS_DEC_IND,
+    META_DUAL_ELGBL_STUS_JAN_CD,
+    META_DUAL_ELGBL_STUS_FEB_CD,
+    META_DUAL_ELGBL_STUS_MAR_CD,
+    META_DUAL_ELGBL_STUS_APR_CD,
+    META_DUAL_ELGBL_STUS_MAY_CD,
+    META_DUAL_ELGBL_STUS_JUN_CD,
+    META_DUAL_ELGBL_STUS_JUL_CD,
+    META_DUAL_ELGBL_STUS_AUG_CD,
+    META_DUAL_ELGBL_STUS_SEPT_CD,
+    META_DUAL_ELGBL_STUS_OCT_CD,
+    META_DUAL_ELGBL_STUS_NOV_CD,
+    META_DUAL_ELGBL_STUS_DEC_CD,
+    CST_SHR_GRP_JAN_CD,
+    CST_SHR_GRP_FEB_CD,
+    CST_SHR_GRP_MAR_CD,
+    CST_SHR_GRP_APR_CD,
+    CST_SHR_GRP_MAY_CD,
+    CST_SHR_GRP_JUN_CD,
+    CST_SHR_GRP_JUL_CD,
+    CST_SHR_GRP_AUG_CD,
+    CST_SHR_GRP_SEPT_CD,
+    CST_SHR_GRP_OCT_CD,
+    CST_SHR_GRP_NOV_CD,
+    CST_SHR_GRP_DEC_CD,
+    DRVD_LINE_1_ADR,
+    DRVD_LINE_2_ADR,
+    DRVD_LINE_3_ADR,
+    DRVD_LINE_4_ADR,
+    DRVD_LINE_5_ADR,
+    DRVD_LINE_6_ADR,
+    CITY_NAME,
+    STATE_CD,
+    STATE_CNTY_ZIP_CD,
+    EFCTV_BGN_DT,
+    EFCTV_END_DT,
+    BENE_LINK_KEY
+  }
+  
+  static final BeneficiaryFields[] beneficiaryPartDContractFields = {
+    BeneficiaryFields.PTD_CNTRCT_JAN_ID,
+    BeneficiaryFields.PTD_CNTRCT_FEB_ID,
+    BeneficiaryFields.PTD_CNTRCT_MAR_ID,
+    BeneficiaryFields.PTD_CNTRCT_APR_ID,
+    BeneficiaryFields.PTD_CNTRCT_MAY_ID,
+    BeneficiaryFields.PTD_CNTRCT_JUN_ID,
+    BeneficiaryFields.PTD_CNTRCT_JUL_ID,
+    BeneficiaryFields.PTD_CNTRCT_AUG_ID,
+    BeneficiaryFields.PTD_CNTRCT_SEPT_ID,
+    BeneficiaryFields.PTD_CNTRCT_OCT_ID,
+    BeneficiaryFields.PTD_CNTRCT_NOV_ID,
+    BeneficiaryFields.PTD_CNTRCT_DEC_ID
+  };
+
+  /* package access */
+  enum BeneficiaryHistoryFields {
+    DML_IND,
+    BENE_ID,
+    STATE_CODE,
+    BENE_COUNTY_CD,
+    BENE_ZIP_CD,
+    BENE_BIRTH_DT,
+    BENE_SEX_IDENT_CD,
+    BENE_RACE_CD,
+    BENE_ENTLMT_RSN_ORIG,
+    BENE_ENTLMT_RSN_CURR,
+    BENE_ESRD_IND,
+    BENE_MDCR_STATUS_CD,
+    BENE_PTA_TRMNTN_CD,
+    BENE_PTB_TRMNTN_CD,
+    BENE_CRNT_HIC_NUM,
+    BENE_SRNM_NAME,
+    BENE_GVN_NAME,
+    BENE_MDL_NAME,
+    MBI_NUM,
+    EFCTV_BGN_DT,
+    EFCTV_END_DT
+  }
+  
+  enum OutpatientFields {
+    DML_IND,
+    BENE_ID,
+    CLM_ID,
+    CLM_GRP_ID,
+    FINAL_ACTION,
+    NCH_NEAR_LINE_REC_IDENT_CD,
+    NCH_CLM_TYPE_CD,
+    CLM_FROM_DT,
+    CLM_THRU_DT,
+    NCH_WKLY_PROC_DT,
+    FI_CLM_PROC_DT,
+    CLAIM_QUERY_CODE,
+    PRVDR_NUM,
+    CLM_FAC_TYPE_CD,
+    CLM_SRVC_CLSFCTN_TYPE_CD,
+    CLM_FREQ_CD,
+    FI_NUM,
+    CLM_MDCR_NON_PMT_RSN_CD,
+    CLM_PMT_AMT,
+    NCH_PRMRY_PYR_CLM_PD_AMT,
+    NCH_PRMRY_PYR_CD,
+    PRVDR_STATE_CD,
+    ORG_NPI_NUM,
+    AT_PHYSN_UPIN,
+    AT_PHYSN_NPI,
+    OP_PHYSN_UPIN,
+    OP_PHYSN_NPI,
+    OT_PHYSN_UPIN,
+    OT_PHYSN_NPI,
+    CLM_MCO_PD_SW,
+    PTNT_DSCHRG_STUS_CD,
+    CLM_TOT_CHRG_AMT,
+    NCH_BENE_BLOOD_DDCTBL_LBLTY_AM,
+    NCH_PROFNL_CMPNT_CHRG_AMT,
+    PRNCPAL_DGNS_CD,
+    PRNCPAL_DGNS_VRSN_CD,
+    ICD_DGNS_CD1,
+    ICD_DGNS_VRSN_CD1,
+    ICD_DGNS_CD2,
+    ICD_DGNS_VRSN_CD2,
+    ICD_DGNS_CD3,
+    ICD_DGNS_VRSN_CD3,
+    ICD_DGNS_CD4,
+    ICD_DGNS_VRSN_CD4,
+    ICD_DGNS_CD5,
+    ICD_DGNS_VRSN_CD5,
+    ICD_DGNS_CD6,
+    ICD_DGNS_VRSN_CD6,
+    ICD_DGNS_CD7,
+    ICD_DGNS_VRSN_CD7,
+    ICD_DGNS_CD8,
+    ICD_DGNS_VRSN_CD8,
+    ICD_DGNS_CD9,
+    ICD_DGNS_VRSN_CD9,
+    ICD_DGNS_CD10,
+    ICD_DGNS_VRSN_CD10,
+    ICD_DGNS_CD11,
+    ICD_DGNS_VRSN_CD11,
+    ICD_DGNS_CD12,
+    ICD_DGNS_VRSN_CD12,
+    ICD_DGNS_CD13,
+    ICD_DGNS_VRSN_CD13,
+    ICD_DGNS_CD14,
+    ICD_DGNS_VRSN_CD14,
+    ICD_DGNS_CD15,
+    ICD_DGNS_VRSN_CD15,
+    ICD_DGNS_CD16,
+    ICD_DGNS_VRSN_CD16,
+    ICD_DGNS_CD17,
+    ICD_DGNS_VRSN_CD17,
+    ICD_DGNS_CD18,
+    ICD_DGNS_VRSN_CD18,
+    ICD_DGNS_CD19,
+    ICD_DGNS_VRSN_CD19,
+    ICD_DGNS_CD20,
+    ICD_DGNS_VRSN_CD20,
+    ICD_DGNS_CD21,
+    ICD_DGNS_VRSN_CD21,
+    ICD_DGNS_CD22,
+    ICD_DGNS_VRSN_CD22,
+    ICD_DGNS_CD23,
+    ICD_DGNS_VRSN_CD23,
+    ICD_DGNS_CD24,
+    ICD_DGNS_VRSN_CD24,
+    ICD_DGNS_CD25,
+    ICD_DGNS_VRSN_CD25,
+    FST_DGNS_E_CD,
+    FST_DGNS_E_VRSN_CD,
+    ICD_DGNS_E_CD1,
+    ICD_DGNS_E_VRSN_CD1,
+    ICD_DGNS_E_CD2,
+    ICD_DGNS_E_VRSN_CD2,
+    ICD_DGNS_E_CD3,
+    ICD_DGNS_E_VRSN_CD3,
+    ICD_DGNS_E_CD4,
+    ICD_DGNS_E_VRSN_CD4,
+    ICD_DGNS_E_CD5,
+    ICD_DGNS_E_VRSN_CD5,
+    ICD_DGNS_E_CD6,
+    ICD_DGNS_E_VRSN_CD6,
+    ICD_DGNS_E_CD7,
+    ICD_DGNS_E_VRSN_CD7,
+    ICD_DGNS_E_CD8,
+    ICD_DGNS_E_VRSN_CD8,
+    ICD_DGNS_E_CD9,
+    ICD_DGNS_E_VRSN_CD9,
+    ICD_DGNS_E_CD10,
+    ICD_DGNS_E_VRSN_CD10,
+    ICD_DGNS_E_CD11,
+    ICD_DGNS_E_VRSN_CD11,
+    ICD_DGNS_E_CD12,
+    ICD_DGNS_E_VRSN_CD12,
+    ICD_PRCDR_CD1,
+    ICD_PRCDR_VRSN_CD1,
+    PRCDR_DT1,
+    ICD_PRCDR_CD2,
+    ICD_PRCDR_VRSN_CD2,
+    PRCDR_DT2,
+    ICD_PRCDR_CD3,
+    ICD_PRCDR_VRSN_CD3,
+    PRCDR_DT3,
+    ICD_PRCDR_CD4,
+    ICD_PRCDR_VRSN_CD4,
+    PRCDR_DT4,
+    ICD_PRCDR_CD5,
+    ICD_PRCDR_VRSN_CD5,
+    PRCDR_DT5,
+    ICD_PRCDR_CD6,
+    ICD_PRCDR_VRSN_CD6,
+    PRCDR_DT6,
+    ICD_PRCDR_CD7,
+    ICD_PRCDR_VRSN_CD7,
+    PRCDR_DT7,
+    ICD_PRCDR_CD8,
+    ICD_PRCDR_VRSN_CD8,
+    PRCDR_DT8,
+    ICD_PRCDR_CD9,
+    ICD_PRCDR_VRSN_CD9,
+    PRCDR_DT9,
+    ICD_PRCDR_CD10,
+    ICD_PRCDR_VRSN_CD10,
+    PRCDR_DT10,
+    ICD_PRCDR_CD11,
+    ICD_PRCDR_VRSN_CD11,
+    PRCDR_DT11,
+    ICD_PRCDR_CD12,
+    ICD_PRCDR_VRSN_CD12,
+    PRCDR_DT12,
+    ICD_PRCDR_CD13,
+    ICD_PRCDR_VRSN_CD13,
+    PRCDR_DT13,
+    ICD_PRCDR_CD14,
+    ICD_PRCDR_VRSN_CD14,
+    PRCDR_DT14,
+    ICD_PRCDR_CD15,
+    ICD_PRCDR_VRSN_CD15,
+    PRCDR_DT15,
+    ICD_PRCDR_CD16,
+    ICD_PRCDR_VRSN_CD16,
+    PRCDR_DT16,
+    ICD_PRCDR_CD17,
+    ICD_PRCDR_VRSN_CD17,
+    PRCDR_DT17,
+    ICD_PRCDR_CD18,
+    ICD_PRCDR_VRSN_CD18,
+    PRCDR_DT18,
+    ICD_PRCDR_CD19,
+    ICD_PRCDR_VRSN_CD19,
+    PRCDR_DT19,
+    ICD_PRCDR_CD20,
+    ICD_PRCDR_VRSN_CD20,
+    PRCDR_DT20,
+    ICD_PRCDR_CD21,
+    ICD_PRCDR_VRSN_CD21,
+    PRCDR_DT21,
+    ICD_PRCDR_CD22,
+    ICD_PRCDR_VRSN_CD22,
+    PRCDR_DT22,
+    ICD_PRCDR_CD23,
+    ICD_PRCDR_VRSN_CD23,
+    PRCDR_DT23,
+    ICD_PRCDR_CD24,
+    ICD_PRCDR_VRSN_CD24,
+    PRCDR_DT24,
+    ICD_PRCDR_CD25,
+    ICD_PRCDR_VRSN_CD25,
+    PRCDR_DT25,
+    RSN_VISIT_CD1,
+    RSN_VISIT_VRSN_CD1,
+    RSN_VISIT_CD2,
+    RSN_VISIT_VRSN_CD2,
+    RSN_VISIT_CD3,
+    RSN_VISIT_VRSN_CD3,
+    NCH_BENE_PTB_DDCTBL_AMT,
+    NCH_BENE_PTB_COINSRNC_AMT,
+    CLM_OP_PRVDR_PMT_AMT,
+    CLM_OP_BENE_PMT_AMT,
+    FI_DOC_CLM_CNTL_NUM,
+    FI_ORIG_CLM_CNTL_NUM,
+    CLM_LINE_NUM,
+    REV_CNTR,
+    REV_CNTR_DT,
+    REV_CNTR_1ST_ANSI_CD,
+    REV_CNTR_2ND_ANSI_CD,
+    REV_CNTR_3RD_ANSI_CD,
+    REV_CNTR_4TH_ANSI_CD,
+    REV_CNTR_APC_HIPPS_CD,
+    HCPCS_CD,
+    HCPCS_1ST_MDFR_CD,
+    HCPCS_2ND_MDFR_CD,
+    REV_CNTR_PMT_MTHD_IND_CD,
+    REV_CNTR_DSCNT_IND_CD,
+    REV_CNTR_PACKG_IND_CD,
+    REV_CNTR_OTAF_PMT_CD,
+    REV_CNTR_IDE_NDC_UPC_NUM,
+    REV_CNTR_UNIT_CNT,
+    REV_CNTR_RATE_AMT,
+    REV_CNTR_BLOOD_DDCTBL_AMT,
+    REV_CNTR_CASH_DDCTBL_AMT,
+    REV_CNTR_COINSRNC_WGE_ADJSTD_C,
+    REV_CNTR_RDCD_COINSRNC_AMT,
+    REV_CNTR_1ST_MSP_PD_AMT,
+    REV_CNTR_2ND_MSP_PD_AMT,
+    REV_CNTR_PRVDR_PMT_AMT,
+    REV_CNTR_BENE_PMT_AMT,
+    REV_CNTR_PTNT_RSPNSBLTY_PMT,
+    REV_CNTR_PMT_AMT_AMT,
+    REV_CNTR_TOT_CHRG_AMT,
+    REV_CNTR_NCVRD_CHRG_AMT,
+    REV_CNTR_STUS_IND_CD,
+    REV_CNTR_NDC_QTY,
+    REV_CNTR_NDC_QTY_QLFR_CD,
+    RNDRNG_PHYSN_UPIN,
+    RNDRNG_PHYSN_NPI
+  }
+
+  /* package access */
+  enum InpatientFields {
+    DML_IND,
+    BENE_ID,
+    CLM_ID,
+    CLM_GRP_ID,
+    FINAL_ACTION,
+    NCH_NEAR_LINE_REC_IDENT_CD,
+    NCH_CLM_TYPE_CD,
+    CLM_FROM_DT,
+    CLM_THRU_DT,
+    NCH_WKLY_PROC_DT,
+    FI_CLM_PROC_DT,
+    CLAIM_QUERY_CODE,
+    PRVDR_NUM,
+    CLM_FAC_TYPE_CD,
+    CLM_SRVC_CLSFCTN_TYPE_CD,
+    CLM_FREQ_CD,
+    FI_NUM,
+    CLM_MDCR_NON_PMT_RSN_CD,
+    CLM_PMT_AMT,
+    NCH_PRMRY_PYR_CLM_PD_AMT,
+    NCH_PRMRY_PYR_CD,
+    FI_CLM_ACTN_CD,
+    PRVDR_STATE_CD,
+    ORG_NPI_NUM,
+    AT_PHYSN_UPIN,
+    AT_PHYSN_NPI,
+    OP_PHYSN_UPIN,
+    OP_PHYSN_NPI,
+    OT_PHYSN_UPIN,
+    OT_PHYSN_NPI,
+    CLM_MCO_PD_SW,
+    PTNT_DSCHRG_STUS_CD,
+    CLM_PPS_IND_CD,
+    CLM_TOT_CHRG_AMT,
+    CLM_ADMSN_DT,
+    CLM_IP_ADMSN_TYPE_CD,
+    CLM_SRC_IP_ADMSN_CD,
+    NCH_PTNT_STATUS_IND_CD,
+    CLM_PASS_THRU_PER_DIEM_AMT,
+    NCH_BENE_IP_DDCTBL_AMT,
+    NCH_BENE_PTA_COINSRNC_LBLTY_AM,
+    NCH_BENE_BLOOD_DDCTBL_LBLTY_AM,
+    NCH_PROFNL_CMPNT_CHRG_AMT,
+    NCH_IP_NCVRD_CHRG_AMT,
+    NCH_IP_TOT_DDCTN_AMT,
+    CLM_TOT_PPS_CPTL_AMT,
+    CLM_PPS_CPTL_FSP_AMT,
+    CLM_PPS_CPTL_OUTLIER_AMT,
+    CLM_PPS_CPTL_DSPRPRTNT_SHR_AMT,
+    CLM_PPS_CPTL_IME_AMT,
+    CLM_PPS_CPTL_EXCPTN_AMT,
+    CLM_PPS_OLD_CPTL_HLD_HRMLS_AMT,
+    CLM_PPS_CPTL_DRG_WT_NUM,
+    CLM_UTLZTN_DAY_CNT,
+    BENE_TOT_COINSRNC_DAYS_CNT,
+    BENE_LRD_USED_CNT,
+    CLM_NON_UTLZTN_DAYS_CNT,
+    NCH_BLOOD_PNTS_FRNSHD_QTY,
+    NCH_VRFD_NCVRD_STAY_FROM_DT,
+    NCH_VRFD_NCVRD_STAY_THRU_DT,
+    NCH_ACTV_OR_CVRD_LVL_CARE_THRU,
+    NCH_BENE_MDCR_BNFTS_EXHTD_DT_I,
+    NCH_BENE_DSCHRG_DT,
+    CLM_DRG_CD,
+    CLM_DRG_OUTLIER_STAY_CD,
+    NCH_DRG_OUTLIER_APRVD_PMT_AMT,
+    ADMTG_DGNS_CD,
+    ADMTG_DGNS_VRSN_CD,
+    PRNCPAL_DGNS_CD,
+    PRNCPAL_DGNS_VRSN_CD,
+    ICD_DGNS_CD1,
+    ICD_DGNS_VRSN_CD1,
+    CLM_POA_IND_SW1,
+    ICD_DGNS_CD2,
+    ICD_DGNS_VRSN_CD2,
+    CLM_POA_IND_SW2,
+    ICD_DGNS_CD3,
+    ICD_DGNS_VRSN_CD3,
+    CLM_POA_IND_SW3,
+    ICD_DGNS_CD4,
+    ICD_DGNS_VRSN_CD4,
+    CLM_POA_IND_SW4,
+    ICD_DGNS_CD5,
+    ICD_DGNS_VRSN_CD5,
+    CLM_POA_IND_SW5,
+    ICD_DGNS_CD6,
+    ICD_DGNS_VRSN_CD6,
+    CLM_POA_IND_SW6,
+    ICD_DGNS_CD7,
+    ICD_DGNS_VRSN_CD7,
+    CLM_POA_IND_SW7,
+    ICD_DGNS_CD8,
+    ICD_DGNS_VRSN_CD8,
+    CLM_POA_IND_SW8,
+    ICD_DGNS_CD9,
+    ICD_DGNS_VRSN_CD9,
+    CLM_POA_IND_SW9,
+    ICD_DGNS_CD10,
+    ICD_DGNS_VRSN_CD10,
+    CLM_POA_IND_SW10,
+    ICD_DGNS_CD11,
+    ICD_DGNS_VRSN_CD11,
+    CLM_POA_IND_SW11,
+    ICD_DGNS_CD12,
+    ICD_DGNS_VRSN_CD12,
+    CLM_POA_IND_SW12,
+    ICD_DGNS_CD13,
+    ICD_DGNS_VRSN_CD13,
+    CLM_POA_IND_SW13,
+    ICD_DGNS_CD14,
+    ICD_DGNS_VRSN_CD14,
+    CLM_POA_IND_SW14,
+    ICD_DGNS_CD15,
+    ICD_DGNS_VRSN_CD15,
+    CLM_POA_IND_SW15,
+    ICD_DGNS_CD16,
+    ICD_DGNS_VRSN_CD16,
+    CLM_POA_IND_SW16,
+    ICD_DGNS_CD17,
+    ICD_DGNS_VRSN_CD17,
+    CLM_POA_IND_SW17,
+    ICD_DGNS_CD18,
+    ICD_DGNS_VRSN_CD18,
+    CLM_POA_IND_SW18,
+    ICD_DGNS_CD19,
+    ICD_DGNS_VRSN_CD19,
+    CLM_POA_IND_SW19,
+    ICD_DGNS_CD20,
+    ICD_DGNS_VRSN_CD20,
+    CLM_POA_IND_SW20,
+    ICD_DGNS_CD21,
+    ICD_DGNS_VRSN_CD21,
+    CLM_POA_IND_SW21,
+    ICD_DGNS_CD22,
+    ICD_DGNS_VRSN_CD22,
+    CLM_POA_IND_SW22,
+    ICD_DGNS_CD23,
+    ICD_DGNS_VRSN_CD23,
+    CLM_POA_IND_SW23,
+    ICD_DGNS_CD24,
+    ICD_DGNS_VRSN_CD24,
+    CLM_POA_IND_SW24,
+    ICD_DGNS_CD25,
+    ICD_DGNS_VRSN_CD25,
+    CLM_POA_IND_SW25,
+    FST_DGNS_E_CD,
+    FST_DGNS_E_VRSN_CD,
+    ICD_DGNS_E_CD1,
+    ICD_DGNS_E_VRSN_CD1,
+    CLM_E_POA_IND_SW1,
+    ICD_DGNS_E_CD2,
+    ICD_DGNS_E_VRSN_CD2,
+    CLM_E_POA_IND_SW2,
+    ICD_DGNS_E_CD3,
+    ICD_DGNS_E_VRSN_CD3,
+    CLM_E_POA_IND_SW3,
+    ICD_DGNS_E_CD4,
+    ICD_DGNS_E_VRSN_CD4,
+    CLM_E_POA_IND_SW4,
+    ICD_DGNS_E_CD5,
+    ICD_DGNS_E_VRSN_CD5,
+    CLM_E_POA_IND_SW5,
+    ICD_DGNS_E_CD6,
+    ICD_DGNS_E_VRSN_CD6,
+    CLM_E_POA_IND_SW6,
+    ICD_DGNS_E_CD7,
+    ICD_DGNS_E_VRSN_CD7,
+    CLM_E_POA_IND_SW7,
+    ICD_DGNS_E_CD8,
+    ICD_DGNS_E_VRSN_CD8,
+    CLM_E_POA_IND_SW8,
+    ICD_DGNS_E_CD9,
+    ICD_DGNS_E_VRSN_CD9,
+    CLM_E_POA_IND_SW9,
+    ICD_DGNS_E_CD10,
+    ICD_DGNS_E_VRSN_CD10,
+    CLM_E_POA_IND_SW10,
+    ICD_DGNS_E_CD11,
+    ICD_DGNS_E_VRSN_CD11,
+    CLM_E_POA_IND_SW11,
+    ICD_DGNS_E_CD12,
+    ICD_DGNS_E_VRSN_CD12,
+    CLM_E_POA_IND_SW12,
+    ICD_PRCDR_CD1,
+    ICD_PRCDR_VRSN_CD1,
+    PRCDR_DT1,
+    ICD_PRCDR_CD2,
+    ICD_PRCDR_VRSN_CD2,
+    PRCDR_DT2,
+    ICD_PRCDR_CD3,
+    ICD_PRCDR_VRSN_CD3,
+    PRCDR_DT3,
+    ICD_PRCDR_CD4,
+    ICD_PRCDR_VRSN_CD4,
+    PRCDR_DT4,
+    ICD_PRCDR_CD5,
+    ICD_PRCDR_VRSN_CD5,
+    PRCDR_DT5,
+    ICD_PRCDR_CD6,
+    ICD_PRCDR_VRSN_CD6,
+    PRCDR_DT6,
+    ICD_PRCDR_CD7,
+    ICD_PRCDR_VRSN_CD7,
+    PRCDR_DT7,
+    ICD_PRCDR_CD8,
+    ICD_PRCDR_VRSN_CD8,
+    PRCDR_DT8,
+    ICD_PRCDR_CD9,
+    ICD_PRCDR_VRSN_CD9,
+    PRCDR_DT9,
+    ICD_PRCDR_CD10,
+    ICD_PRCDR_VRSN_CD10,
+    PRCDR_DT10,
+    ICD_PRCDR_CD11,
+    ICD_PRCDR_VRSN_CD11,
+    PRCDR_DT11,
+    ICD_PRCDR_CD12,
+    ICD_PRCDR_VRSN_CD12,
+    PRCDR_DT12,
+    ICD_PRCDR_CD13,
+    ICD_PRCDR_VRSN_CD13,
+    PRCDR_DT13,
+    ICD_PRCDR_CD14,
+    ICD_PRCDR_VRSN_CD14,
+    PRCDR_DT14,
+    ICD_PRCDR_CD15,
+    ICD_PRCDR_VRSN_CD15,
+    PRCDR_DT15,
+    ICD_PRCDR_CD16,
+    ICD_PRCDR_VRSN_CD16,
+    PRCDR_DT16,
+    ICD_PRCDR_CD17,
+    ICD_PRCDR_VRSN_CD17,
+    PRCDR_DT17,
+    ICD_PRCDR_CD18,
+    ICD_PRCDR_VRSN_CD18,
+    PRCDR_DT18,
+    ICD_PRCDR_CD19,
+    ICD_PRCDR_VRSN_CD19,
+    PRCDR_DT19,
+    ICD_PRCDR_CD20,
+    ICD_PRCDR_VRSN_CD20,
+    PRCDR_DT20,
+    ICD_PRCDR_CD21,
+    ICD_PRCDR_VRSN_CD21,
+    PRCDR_DT21,
+    ICD_PRCDR_CD22,
+    ICD_PRCDR_VRSN_CD22,
+    PRCDR_DT22,
+    ICD_PRCDR_CD23,
+    ICD_PRCDR_VRSN_CD23,
+    PRCDR_DT23,
+    ICD_PRCDR_CD24,
+    ICD_PRCDR_VRSN_CD24,
+    PRCDR_DT24,
+    ICD_PRCDR_CD25,
+    ICD_PRCDR_VRSN_CD25,
+    PRCDR_DT25,
+    IME_OP_CLM_VAL_AMT,
+    DSH_OP_CLM_VAL_AMT,
+    CLM_UNCOMPD_CARE_PMT_AMT,
+    FI_DOC_CLM_CNTL_NUM,
+    FI_ORIG_CLM_CNTL_NUM,
+    CLM_LINE_NUM,
+    REV_CNTR,
+    HCPCS_CD,
+    REV_CNTR_UNIT_CNT,
+    REV_CNTR_RATE_AMT,
+    REV_CNTR_TOT_CHRG_AMT,
+    REV_CNTR_NCVRD_CHRG_AMT,
+    REV_CNTR_DDCTBL_COINSRNC_CD,
+    REV_CNTR_NDC_QTY,
+    REV_CNTR_NDC_QTY_QLFR_CD,
+    RNDRNG_PHYSN_UPIN,
+    RNDRNG_PHYSN_NPI
+  }
+
+  static final InpatientFields[][] inpatientDxFields = {
+    { InpatientFields.ICD_DGNS_CD1, InpatientFields.ICD_DGNS_VRSN_CD1,
+      InpatientFields.CLM_POA_IND_SW1 },
+    { InpatientFields.ICD_DGNS_CD2, InpatientFields.ICD_DGNS_VRSN_CD2,
+      InpatientFields.CLM_POA_IND_SW2 },
+    { InpatientFields.ICD_DGNS_CD3, InpatientFields.ICD_DGNS_VRSN_CD3,
+      InpatientFields.CLM_POA_IND_SW3 },
+    { InpatientFields.ICD_DGNS_CD4, InpatientFields.ICD_DGNS_VRSN_CD4,
+      InpatientFields.CLM_POA_IND_SW4 },
+    { InpatientFields.ICD_DGNS_CD5, InpatientFields.ICD_DGNS_VRSN_CD5,
+      InpatientFields.CLM_POA_IND_SW5 },
+    { InpatientFields.ICD_DGNS_CD6, InpatientFields.ICD_DGNS_VRSN_CD6,
+      InpatientFields.CLM_POA_IND_SW6 },
+    { InpatientFields.ICD_DGNS_CD7, InpatientFields.ICD_DGNS_VRSN_CD7,
+      InpatientFields.CLM_POA_IND_SW7 },
+    { InpatientFields.ICD_DGNS_CD8, InpatientFields.ICD_DGNS_VRSN_CD8,
+      InpatientFields.CLM_POA_IND_SW8 },
+    { InpatientFields.ICD_DGNS_CD9, InpatientFields.ICD_DGNS_VRSN_CD9,
+      InpatientFields.CLM_POA_IND_SW9 },
+    { InpatientFields.ICD_DGNS_CD10, InpatientFields.ICD_DGNS_VRSN_CD10,
+      InpatientFields.CLM_POA_IND_SW10 },
+    { InpatientFields.ICD_DGNS_CD11, InpatientFields.ICD_DGNS_VRSN_CD11,
+      InpatientFields.CLM_POA_IND_SW11 },
+    { InpatientFields.ICD_DGNS_CD12, InpatientFields.ICD_DGNS_VRSN_CD12,
+      InpatientFields.CLM_POA_IND_SW12 },
+    { InpatientFields.ICD_DGNS_CD13, InpatientFields.ICD_DGNS_VRSN_CD13,
+      InpatientFields.CLM_POA_IND_SW13 },
+    { InpatientFields.ICD_DGNS_CD14, InpatientFields.ICD_DGNS_VRSN_CD14,
+      InpatientFields.CLM_POA_IND_SW14 },
+    { InpatientFields.ICD_DGNS_CD15, InpatientFields.ICD_DGNS_VRSN_CD15,
+      InpatientFields.CLM_POA_IND_SW15 },
+    { InpatientFields.ICD_DGNS_CD16, InpatientFields.ICD_DGNS_VRSN_CD16,
+      InpatientFields.CLM_POA_IND_SW16 },
+    { InpatientFields.ICD_DGNS_CD17, InpatientFields.ICD_DGNS_VRSN_CD17,
+      InpatientFields.CLM_POA_IND_SW17 },
+    { InpatientFields.ICD_DGNS_CD18, InpatientFields.ICD_DGNS_VRSN_CD18,
+      InpatientFields.CLM_POA_IND_SW18 },
+    { InpatientFields.ICD_DGNS_CD19, InpatientFields.ICD_DGNS_VRSN_CD19,
+      InpatientFields.CLM_POA_IND_SW19 },
+    { InpatientFields.ICD_DGNS_CD20, InpatientFields.ICD_DGNS_VRSN_CD20,
+      InpatientFields.CLM_POA_IND_SW20 },
+    { InpatientFields.ICD_DGNS_CD21, InpatientFields.ICD_DGNS_VRSN_CD21,
+      InpatientFields.CLM_POA_IND_SW21 },
+    { InpatientFields.ICD_DGNS_CD22, InpatientFields.ICD_DGNS_VRSN_CD22,
+      InpatientFields.CLM_POA_IND_SW22 },
+    { InpatientFields.ICD_DGNS_CD23, InpatientFields.ICD_DGNS_VRSN_CD23,
+      InpatientFields.CLM_POA_IND_SW23 },
+    { InpatientFields.ICD_DGNS_CD24, InpatientFields.ICD_DGNS_VRSN_CD24,
+      InpatientFields.CLM_POA_IND_SW24 },
+    { InpatientFields.ICD_DGNS_CD25, InpatientFields.ICD_DGNS_VRSN_CD25,
+      InpatientFields.CLM_POA_IND_SW25 }
+  };
+
+  static final InpatientFields[][] inpatientPxFields = {
+    { InpatientFields.ICD_PRCDR_CD1, InpatientFields.ICD_PRCDR_VRSN_CD1,
+      InpatientFields.PRCDR_DT1 },
+    { InpatientFields.ICD_PRCDR_CD2, InpatientFields.ICD_PRCDR_VRSN_CD2,
+      InpatientFields.PRCDR_DT2 },
+    { InpatientFields.ICD_PRCDR_CD3, InpatientFields.ICD_PRCDR_VRSN_CD3,
+      InpatientFields.PRCDR_DT3 },
+    { InpatientFields.ICD_PRCDR_CD4, InpatientFields.ICD_PRCDR_VRSN_CD4,
+      InpatientFields.PRCDR_DT4 },
+    { InpatientFields.ICD_PRCDR_CD5, InpatientFields.ICD_PRCDR_VRSN_CD5,
+      InpatientFields.PRCDR_DT5 },
+    { InpatientFields.ICD_PRCDR_CD6, InpatientFields.ICD_PRCDR_VRSN_CD6,
+      InpatientFields.PRCDR_DT6 },
+    { InpatientFields.ICD_PRCDR_CD7, InpatientFields.ICD_PRCDR_VRSN_CD7,
+      InpatientFields.PRCDR_DT7 },
+    { InpatientFields.ICD_PRCDR_CD8, InpatientFields.ICD_PRCDR_VRSN_CD8,
+      InpatientFields.PRCDR_DT8 },
+    { InpatientFields.ICD_PRCDR_CD9, InpatientFields.ICD_PRCDR_VRSN_CD9,
+      InpatientFields.PRCDR_DT9 },
+    { InpatientFields.ICD_PRCDR_CD10, InpatientFields.ICD_PRCDR_VRSN_CD10,
+      InpatientFields.PRCDR_DT10 },
+    { InpatientFields.ICD_PRCDR_CD11, InpatientFields.ICD_PRCDR_VRSN_CD11,
+      InpatientFields.PRCDR_DT11 },
+    { InpatientFields.ICD_PRCDR_CD12, InpatientFields.ICD_PRCDR_VRSN_CD12,
+      InpatientFields.PRCDR_DT12 },
+    { InpatientFields.ICD_PRCDR_CD13, InpatientFields.ICD_PRCDR_VRSN_CD13,
+      InpatientFields.PRCDR_DT13 },
+    { InpatientFields.ICD_PRCDR_CD14, InpatientFields.ICD_PRCDR_VRSN_CD14,
+      InpatientFields.PRCDR_DT14 },
+    { InpatientFields.ICD_PRCDR_CD15, InpatientFields.ICD_PRCDR_VRSN_CD15,
+      InpatientFields.PRCDR_DT15 },
+    { InpatientFields.ICD_PRCDR_CD16, InpatientFields.ICD_PRCDR_VRSN_CD16,
+      InpatientFields.PRCDR_DT16 },
+    { InpatientFields.ICD_PRCDR_CD17, InpatientFields.ICD_PRCDR_VRSN_CD17,
+      InpatientFields.PRCDR_DT17 },
+    { InpatientFields.ICD_PRCDR_CD18, InpatientFields.ICD_PRCDR_VRSN_CD18,
+      InpatientFields.PRCDR_DT18 },
+    { InpatientFields.ICD_PRCDR_CD19, InpatientFields.ICD_PRCDR_VRSN_CD19,
+      InpatientFields.PRCDR_DT19 },
+    { InpatientFields.ICD_PRCDR_CD20, InpatientFields.ICD_PRCDR_VRSN_CD20,
+      InpatientFields.PRCDR_DT20 },
+    { InpatientFields.ICD_PRCDR_CD21, InpatientFields.ICD_PRCDR_VRSN_CD21,
+      InpatientFields.PRCDR_DT21 },
+    { InpatientFields.ICD_PRCDR_CD22, InpatientFields.ICD_PRCDR_VRSN_CD22,
+      InpatientFields.PRCDR_DT22 },
+    { InpatientFields.ICD_PRCDR_CD23, InpatientFields.ICD_PRCDR_VRSN_CD23,
+      InpatientFields.PRCDR_DT23 },
+    { InpatientFields.ICD_PRCDR_CD24, InpatientFields.ICD_PRCDR_VRSN_CD24,
+      InpatientFields.PRCDR_DT24 },
+    { InpatientFields.ICD_PRCDR_CD25, InpatientFields.ICD_PRCDR_VRSN_CD25,
+      InpatientFields.PRCDR_DT25 }
+  };
+
+  static final OutpatientFields[][] outpatientDxFields = {
+    { OutpatientFields.ICD_DGNS_CD1, OutpatientFields.ICD_DGNS_VRSN_CD1 },
+    { OutpatientFields.ICD_DGNS_CD2, OutpatientFields.ICD_DGNS_VRSN_CD2 },
+    { OutpatientFields.ICD_DGNS_CD3, OutpatientFields.ICD_DGNS_VRSN_CD3 },
+    { OutpatientFields.ICD_DGNS_CD4, OutpatientFields.ICD_DGNS_VRSN_CD4 },
+    { OutpatientFields.ICD_DGNS_CD5, OutpatientFields.ICD_DGNS_VRSN_CD5 },
+    { OutpatientFields.ICD_DGNS_CD6, OutpatientFields.ICD_DGNS_VRSN_CD6 },
+    { OutpatientFields.ICD_DGNS_CD7, OutpatientFields.ICD_DGNS_VRSN_CD7 },
+    { OutpatientFields.ICD_DGNS_CD8, OutpatientFields.ICD_DGNS_VRSN_CD8 },
+    { OutpatientFields.ICD_DGNS_CD9, OutpatientFields.ICD_DGNS_VRSN_CD9 },
+    { OutpatientFields.ICD_DGNS_CD10, OutpatientFields.ICD_DGNS_VRSN_CD10 },
+    { OutpatientFields.ICD_DGNS_CD11, OutpatientFields.ICD_DGNS_VRSN_CD11 },
+    { OutpatientFields.ICD_DGNS_CD12, OutpatientFields.ICD_DGNS_VRSN_CD12 },
+    { OutpatientFields.ICD_DGNS_CD13, OutpatientFields.ICD_DGNS_VRSN_CD13 },
+    { OutpatientFields.ICD_DGNS_CD14, OutpatientFields.ICD_DGNS_VRSN_CD14 },
+    { OutpatientFields.ICD_DGNS_CD15, OutpatientFields.ICD_DGNS_VRSN_CD15 },
+    { OutpatientFields.ICD_DGNS_CD16, OutpatientFields.ICD_DGNS_VRSN_CD16 },
+    { OutpatientFields.ICD_DGNS_CD17, OutpatientFields.ICD_DGNS_VRSN_CD17 },
+    { OutpatientFields.ICD_DGNS_CD18, OutpatientFields.ICD_DGNS_VRSN_CD18 },
+    { OutpatientFields.ICD_DGNS_CD19, OutpatientFields.ICD_DGNS_VRSN_CD19 },
+    { OutpatientFields.ICD_DGNS_CD20, OutpatientFields.ICD_DGNS_VRSN_CD20 },
+    { OutpatientFields.ICD_DGNS_CD21, OutpatientFields.ICD_DGNS_VRSN_CD21 },
+    { OutpatientFields.ICD_DGNS_CD22, OutpatientFields.ICD_DGNS_VRSN_CD22 },
+    { OutpatientFields.ICD_DGNS_CD23, OutpatientFields.ICD_DGNS_VRSN_CD23 },
+    { OutpatientFields.ICD_DGNS_CD24, OutpatientFields.ICD_DGNS_VRSN_CD24 },
+    { OutpatientFields.ICD_DGNS_CD25, OutpatientFields.ICD_DGNS_VRSN_CD25 }
+  };
+
+  static final OutpatientFields[][] outpatientPxFields = {
+    { OutpatientFields.ICD_PRCDR_CD1, OutpatientFields.ICD_PRCDR_VRSN_CD1,
+      OutpatientFields.PRCDR_DT1 },
+    { OutpatientFields.ICD_PRCDR_CD2, OutpatientFields.ICD_PRCDR_VRSN_CD2,
+      OutpatientFields.PRCDR_DT2 },
+    { OutpatientFields.ICD_PRCDR_CD3, OutpatientFields.ICD_PRCDR_VRSN_CD3,
+      OutpatientFields.PRCDR_DT3 },
+    { OutpatientFields.ICD_PRCDR_CD4, OutpatientFields.ICD_PRCDR_VRSN_CD4,
+      OutpatientFields.PRCDR_DT4 },
+    { OutpatientFields.ICD_PRCDR_CD5, OutpatientFields.ICD_PRCDR_VRSN_CD5,
+      OutpatientFields.PRCDR_DT5 },
+    { OutpatientFields.ICD_PRCDR_CD6, OutpatientFields.ICD_PRCDR_VRSN_CD6,
+      OutpatientFields.PRCDR_DT6 },
+    { OutpatientFields.ICD_PRCDR_CD7, OutpatientFields.ICD_PRCDR_VRSN_CD7,
+      OutpatientFields.PRCDR_DT7 },
+    { OutpatientFields.ICD_PRCDR_CD8, OutpatientFields.ICD_PRCDR_VRSN_CD8,
+      OutpatientFields.PRCDR_DT8 },
+    { OutpatientFields.ICD_PRCDR_CD9, OutpatientFields.ICD_PRCDR_VRSN_CD9,
+      OutpatientFields.PRCDR_DT9 },
+    { OutpatientFields.ICD_PRCDR_CD10, OutpatientFields.ICD_PRCDR_VRSN_CD10,
+      OutpatientFields.PRCDR_DT10 },
+    { OutpatientFields.ICD_PRCDR_CD11, OutpatientFields.ICD_PRCDR_VRSN_CD11,
+      OutpatientFields.PRCDR_DT11 },
+    { OutpatientFields.ICD_PRCDR_CD12, OutpatientFields.ICD_PRCDR_VRSN_CD12,
+      OutpatientFields.PRCDR_DT12 },
+    { OutpatientFields.ICD_PRCDR_CD13, OutpatientFields.ICD_PRCDR_VRSN_CD13,
+      OutpatientFields.PRCDR_DT13 },
+    { OutpatientFields.ICD_PRCDR_CD14, OutpatientFields.ICD_PRCDR_VRSN_CD14,
+      OutpatientFields.PRCDR_DT14 },
+    { OutpatientFields.ICD_PRCDR_CD15, OutpatientFields.ICD_PRCDR_VRSN_CD15,
+      OutpatientFields.PRCDR_DT15 },
+    { OutpatientFields.ICD_PRCDR_CD16, OutpatientFields.ICD_PRCDR_VRSN_CD16,
+      OutpatientFields.PRCDR_DT16 },
+    { OutpatientFields.ICD_PRCDR_CD17, OutpatientFields.ICD_PRCDR_VRSN_CD17,
+      OutpatientFields.PRCDR_DT17 },
+    { OutpatientFields.ICD_PRCDR_CD18, OutpatientFields.ICD_PRCDR_VRSN_CD18,
+      OutpatientFields.PRCDR_DT18 },
+    { OutpatientFields.ICD_PRCDR_CD19, OutpatientFields.ICD_PRCDR_VRSN_CD19,
+      OutpatientFields.PRCDR_DT19 },
+    { OutpatientFields.ICD_PRCDR_CD20, OutpatientFields.ICD_PRCDR_VRSN_CD20,
+      OutpatientFields.PRCDR_DT20 },
+    { OutpatientFields.ICD_PRCDR_CD21, OutpatientFields.ICD_PRCDR_VRSN_CD21,
+      OutpatientFields.PRCDR_DT21 },
+    { OutpatientFields.ICD_PRCDR_CD22, OutpatientFields.ICD_PRCDR_VRSN_CD22,
+      OutpatientFields.PRCDR_DT22 },
+    { OutpatientFields.ICD_PRCDR_CD23, OutpatientFields.ICD_PRCDR_VRSN_CD23,
+      OutpatientFields.PRCDR_DT23 },
+    { OutpatientFields.ICD_PRCDR_CD24, OutpatientFields.ICD_PRCDR_VRSN_CD24,
+      OutpatientFields.PRCDR_DT24 },
+    { OutpatientFields.ICD_PRCDR_CD25, OutpatientFields.ICD_PRCDR_VRSN_CD25,
+      OutpatientFields.PRCDR_DT25 }
+  };
+
+  /* package access */
+  enum CarrierFields {
+    DML_IND,
+    BENE_ID,
+    CLM_ID,
+    CLM_GRP_ID,
+    FINAL_ACTION,
+    NCH_NEAR_LINE_REC_IDENT_CD,
+    NCH_CLM_TYPE_CD,
+    CLM_FROM_DT,
+    CLM_THRU_DT,
+    NCH_WKLY_PROC_DT,
+    CARR_CLM_ENTRY_CD,
+    CLM_DISP_CD,
+    CARR_NUM,
+    CARR_CLM_PMT_DNL_CD,
+    CLM_PMT_AMT,
+    CARR_CLM_PRMRY_PYR_PD_AMT,
+    RFR_PHYSN_UPIN,
+    RFR_PHYSN_NPI,
+    CARR_CLM_PRVDR_ASGNMT_IND_SW,
+    NCH_CLM_PRVDR_PMT_AMT,
+    NCH_CLM_BENE_PMT_AMT,
+    NCH_CARR_CLM_SBMTD_CHRG_AMT,
+    NCH_CARR_CLM_ALOWD_AMT,
+    CARR_CLM_CASH_DDCTBL_APLD_AMT,
+    CARR_CLM_HCPCS_YR_CD,
+    CARR_CLM_RFRNG_PIN_NUM,
+    PRNCPAL_DGNS_CD,
+    PRNCPAL_DGNS_VRSN_CD,
+    ICD_DGNS_CD1,
+    ICD_DGNS_VRSN_CD1,
+    ICD_DGNS_CD2,
+    ICD_DGNS_VRSN_CD2,
+    ICD_DGNS_CD3,
+    ICD_DGNS_VRSN_CD3,
+    ICD_DGNS_CD4,
+    ICD_DGNS_VRSN_CD4,
+    ICD_DGNS_CD5,
+    ICD_DGNS_VRSN_CD5,
+    ICD_DGNS_CD6,
+    ICD_DGNS_VRSN_CD6,
+    ICD_DGNS_CD7,
+    ICD_DGNS_VRSN_CD7,
+    ICD_DGNS_CD8,
+    ICD_DGNS_VRSN_CD8,
+    ICD_DGNS_CD9,
+    ICD_DGNS_VRSN_CD9,
+    ICD_DGNS_CD10,
+    ICD_DGNS_VRSN_CD10,
+    ICD_DGNS_CD11,
+    ICD_DGNS_VRSN_CD11,
+    ICD_DGNS_CD12,
+    ICD_DGNS_VRSN_CD12,
+    CLM_CLNCL_TRIL_NUM,
+    CARR_CLM_CNTL_NUM,
+    LINE_NUM,
+    CARR_PRFRNG_PIN_NUM,
+    PRF_PHYSN_UPIN,
+    PRF_PHYSN_NPI,
+    ORG_NPI_NUM,
+    CARR_LINE_PRVDR_TYPE_CD,
+    TAX_NUM,
+    PRVDR_STATE_CD,
+    PRVDR_ZIP,
+    PRVDR_SPCLTY,
+    PRTCPTNG_IND_CD,
+    CARR_LINE_RDCD_PMT_PHYS_ASTN_C,
+    LINE_SRVC_CNT,
+    LINE_CMS_TYPE_SRVC_CD,
+    LINE_PLACE_OF_SRVC_CD,
+    CARR_LINE_PRCNG_LCLTY_CD,
+    LINE_1ST_EXPNS_DT,
+    LINE_LAST_EXPNS_DT,
+    HCPCS_CD,
+    HCPCS_1ST_MDFR_CD,
+    HCPCS_2ND_MDFR_CD,
+    BETOS_CD,
+    LINE_NCH_PMT_AMT,
+    LINE_BENE_PMT_AMT,
+    LINE_PRVDR_PMT_AMT,
+    LINE_BENE_PTB_DDCTBL_AMT,
+    LINE_BENE_PRMRY_PYR_CD,
+    LINE_BENE_PRMRY_PYR_PD_AMT,
+    LINE_COINSRNC_AMT,
+    LINE_SBMTD_CHRG_AMT,
+    LINE_ALOWD_CHRG_AMT,
+    LINE_PRCSG_IND_CD,
+    LINE_PMT_80_100_CD,
+    LINE_SERVICE_DEDUCTIBLE,
+    CARR_LINE_MTUS_CNT,
+    CARR_LINE_MTUS_CD,
+    LINE_ICD_DGNS_CD,
+    LINE_ICD_DGNS_VRSN_CD,
+    HPSA_SCRCTY_IND_CD,
+    CARR_LINE_RX_NUM,
+    LINE_HCT_HGB_RSLT_NUM,
+    LINE_HCT_HGB_TYPE_CD,
+    LINE_NDC_CD,
+    CARR_LINE_CLIA_LAB_NUM,
+    CARR_LINE_ANSTHSA_UNIT_CNT
+  }
+
+  static final CarrierFields[][] carrierDxFields = {
+    { CarrierFields.ICD_DGNS_CD1, CarrierFields.ICD_DGNS_VRSN_CD1 },
+    { CarrierFields.ICD_DGNS_CD2, CarrierFields.ICD_DGNS_VRSN_CD2 },
+    { CarrierFields.ICD_DGNS_CD3, CarrierFields.ICD_DGNS_VRSN_CD3 },
+    { CarrierFields.ICD_DGNS_CD4, CarrierFields.ICD_DGNS_VRSN_CD4 },
+    { CarrierFields.ICD_DGNS_CD5, CarrierFields.ICD_DGNS_VRSN_CD5 },
+    { CarrierFields.ICD_DGNS_CD6, CarrierFields.ICD_DGNS_VRSN_CD6 },
+    { CarrierFields.ICD_DGNS_CD7, CarrierFields.ICD_DGNS_VRSN_CD7 },
+    { CarrierFields.ICD_DGNS_CD8, CarrierFields.ICD_DGNS_VRSN_CD8 },
+    { CarrierFields.ICD_DGNS_CD9, CarrierFields.ICD_DGNS_VRSN_CD9 },
+    { CarrierFields.ICD_DGNS_CD10, CarrierFields.ICD_DGNS_VRSN_CD10 },
+    { CarrierFields.ICD_DGNS_CD11, CarrierFields.ICD_DGNS_VRSN_CD11 },
+    { CarrierFields.ICD_DGNS_CD12, CarrierFields.ICD_DGNS_VRSN_CD12 }
+  };
+
+  enum PrescriptionFields {
+    DML_IND,
+    PDE_ID,
+    CLM_GRP_ID,
+    FINAL_ACTION,
+    BENE_ID,
+    SRVC_DT,
+    PD_DT,
+    SRVC_PRVDR_ID_QLFYR_CD,
+    SRVC_PRVDR_ID,
+    PRSCRBR_ID_QLFYR_CD,
+    PRSCRBR_ID,
+    RX_SRVC_RFRNC_NUM,
+    PROD_SRVC_ID,
+    PLAN_CNTRCT_REC_ID,
+    PLAN_PBP_REC_NUM,
+    CMPND_CD,
+    DAW_PROD_SLCTN_CD,
+    QTY_DSPNSD_NUM,
+    DAYS_SUPLY_NUM,
+    FILL_NUM,
+    DSPNSNG_STUS_CD,
+    DRUG_CVRG_STUS_CD,
+    ADJSTMT_DLTN_CD,
+    NSTD_FRMT_CD,
+    PRCNG_EXCPTN_CD,
+    CTSTRPHC_CVRG_CD,
+    GDC_BLW_OOPT_AMT,
+    GDC_ABV_OOPT_AMT,
+    PTNT_PAY_AMT,
+    OTHR_TROOP_AMT,
+    LICS_AMT,
+    PLRO_AMT,
+    CVRD_D_PLAN_PD_AMT,
+    NCVRD_PLAN_PD_AMT,
+    TOT_RX_CST_AMT,
+    RX_ORGN_CD,
+    RPTD_GAP_DSCNT_NUM,
+    BRND_GNRC_CD,
+    PHRMCY_SRVC_TYPE_CD,
+    PTNT_RSDNC_CD,
+    SUBMSN_CLR_CD
+  }
+
+  enum DMEFields {
+    DML_IND,
+    BENE_ID,
+    CLM_ID,
+    CLM_GRP_ID,
+    FINAL_ACTION,
+    NCH_NEAR_LINE_REC_IDENT_CD,
+    NCH_CLM_TYPE_CD,
+    CLM_FROM_DT,
+    CLM_THRU_DT,
+    NCH_WKLY_PROC_DT,
+    CARR_CLM_ENTRY_CD,
+    CLM_DISP_CD,
+    CARR_NUM,
+    CARR_CLM_PMT_DNL_CD,
+    CLM_PMT_AMT,
+    CARR_CLM_PRMRY_PYR_PD_AMT,
+    CARR_CLM_PRVDR_ASGNMT_IND_SW,
+    NCH_CLM_PRVDR_PMT_AMT,
+    NCH_CLM_BENE_PMT_AMT,
+    NCH_CARR_CLM_SBMTD_CHRG_AMT,
+    NCH_CARR_CLM_ALOWD_AMT,
+    CARR_CLM_CASH_DDCTBL_APLD_AMT,
+    CARR_CLM_HCPCS_YR_CD,
+    PRNCPAL_DGNS_CD,
+    PRNCPAL_DGNS_VRSN_CD,
+    ICD_DGNS_CD1,
+    ICD_DGNS_VRSN_CD1,
+    ICD_DGNS_CD2,
+    ICD_DGNS_VRSN_CD2,
+    ICD_DGNS_CD3,
+    ICD_DGNS_VRSN_CD3,
+    ICD_DGNS_CD4,
+    ICD_DGNS_VRSN_CD4,
+    ICD_DGNS_CD5,
+    ICD_DGNS_VRSN_CD5,
+    ICD_DGNS_CD6,
+    ICD_DGNS_VRSN_CD6,
+    ICD_DGNS_CD7,
+    ICD_DGNS_VRSN_CD7,
+    ICD_DGNS_CD8,
+    ICD_DGNS_VRSN_CD8,
+    ICD_DGNS_CD9,
+    ICD_DGNS_VRSN_CD9,
+    ICD_DGNS_CD10,
+    ICD_DGNS_VRSN_CD10,
+    ICD_DGNS_CD11,
+    ICD_DGNS_VRSN_CD11,
+    ICD_DGNS_CD12,
+    ICD_DGNS_VRSN_CD12,
+    RFR_PHYSN_UPIN,
+    RFR_PHYSN_NPI,
+    CLM_CLNCL_TRIL_NUM,
+    CARR_CLM_CNTL_NUM,
+    LINE_NUM,
+    TAX_NUM,
+    PRVDR_SPCLTY,
+    PRTCPTNG_IND_CD,
+    LINE_SRVC_CNT,
+    LINE_CMS_TYPE_SRVC_CD,
+    LINE_PLACE_OF_SRVC_CD,
+    LINE_1ST_EXPNS_DT,
+    LINE_LAST_EXPNS_DT,
+    HCPCS_CD,
+    HCPCS_1ST_MDFR_CD,
+    HCPCS_2ND_MDFR_CD,
+    BETOS_CD,
+    LINE_NCH_PMT_AMT,
+    LINE_BENE_PMT_AMT,
+    LINE_PRVDR_PMT_AMT,
+    LINE_BENE_PTB_DDCTBL_AMT,
+    LINE_BENE_PRMRY_PYR_CD,
+    LINE_BENE_PRMRY_PYR_PD_AMT,
+    LINE_COINSRNC_AMT,
+    LINE_PRMRY_ALOWD_CHRG_AMT,
+    LINE_SBMTD_CHRG_AMT,
+    LINE_ALOWD_CHRG_AMT,
+    LINE_PRCSG_IND_CD,
+    LINE_PMT_80_100_CD,
+    LINE_SERVICE_DEDUCTIBLE,
+    LINE_ICD_DGNS_CD,
+    LINE_ICD_DGNS_VRSN_CD,
+    LINE_DME_PRCHS_PRICE_AMT,
+    PRVDR_NUM,
+    PRVDR_NPI,
+    DMERC_LINE_PRCNG_STATE_CD,
+    PRVDR_STATE_CD,
+    DMERC_LINE_SUPPLR_TYPE_CD,
+    HCPCS_3RD_MDFR_CD,
+    HCPCS_4TH_MDFR_CD,
+    DMERC_LINE_SCRN_SVGS_AMT,
+    DMERC_LINE_MTUS_CNT,
+    DMERC_LINE_MTUS_CD,
+    LINE_HCT_HGB_RSLT_NUM,
+    LINE_HCT_HGB_TYPE_CD,
+    LINE_NDC_CD
+  }
+
+  static final DMEFields[][] dmeDxFields = {
+    { DMEFields.ICD_DGNS_CD1, DMEFields.ICD_DGNS_VRSN_CD1 },
+    { DMEFields.ICD_DGNS_CD2, DMEFields.ICD_DGNS_VRSN_CD2 },
+    { DMEFields.ICD_DGNS_CD3, DMEFields.ICD_DGNS_VRSN_CD3 },
+    { DMEFields.ICD_DGNS_CD4, DMEFields.ICD_DGNS_VRSN_CD4 },
+    { DMEFields.ICD_DGNS_CD5, DMEFields.ICD_DGNS_VRSN_CD5 },
+    { DMEFields.ICD_DGNS_CD6, DMEFields.ICD_DGNS_VRSN_CD6 },
+    { DMEFields.ICD_DGNS_CD7, DMEFields.ICD_DGNS_VRSN_CD7 },
+    { DMEFields.ICD_DGNS_CD8, DMEFields.ICD_DGNS_VRSN_CD8 },
+    { DMEFields.ICD_DGNS_CD9, DMEFields.ICD_DGNS_VRSN_CD9 },
+    { DMEFields.ICD_DGNS_CD10, DMEFields.ICD_DGNS_VRSN_CD10 },
+    { DMEFields.ICD_DGNS_CD11, DMEFields.ICD_DGNS_VRSN_CD11 },
+    { DMEFields.ICD_DGNS_CD12, DMEFields.ICD_DGNS_VRSN_CD12 }
+  };
+
+  enum NPIFields {
+    NPI,
+    ENTITY_TYPE_CODE,
+    REPLACEMENT_NPI,
+    EIN,
+    ORG_NAME,
+    LAST_NAME,
+    FIRST_NAME,
+    MIDDLE_NAME,
+    PREFIX,
+    SUFFIX,
+    CREDENTIALS
+  }
+  
+  enum HHAFields {
+    DML_IND,
+    BENE_ID,
+    CLM_ID,
+    CLM_GRP_ID,
+    FINAL_ACTION,
+    NCH_NEAR_LINE_REC_IDENT_CD,
+    NCH_CLM_TYPE_CD,
+    CLM_FROM_DT,
+    CLM_THRU_DT,
+    NCH_WKLY_PROC_DT,
+    FI_CLM_PROC_DT,
+    PRVDR_NUM,
+    CLM_FAC_TYPE_CD,
+    CLM_SRVC_CLSFCTN_TYPE_CD,
+    CLM_FREQ_CD,
+    FI_NUM,
+    CLM_MDCR_NON_PMT_RSN_CD,
+    CLM_PMT_AMT,
+    NCH_PRMRY_PYR_CLM_PD_AMT,
+    NCH_PRMRY_PYR_CD,
+    PRVDR_STATE_CD,
+    ORG_NPI_NUM,
+    AT_PHYSN_UPIN,
+    AT_PHYSN_NPI,
+    PTNT_DSCHRG_STUS_CD,
+    CLM_PPS_IND_CD,
+    CLM_TOT_CHRG_AMT,
+    PRNCPAL_DGNS_CD,
+    PRNCPAL_DGNS_VRSN_CD,
+    ICD_DGNS_CD1,
+    ICD_DGNS_VRSN_CD1,
+    ICD_DGNS_CD2,
+    ICD_DGNS_VRSN_CD2,
+    ICD_DGNS_CD3,
+    ICD_DGNS_VRSN_CD3,
+    ICD_DGNS_CD4,
+    ICD_DGNS_VRSN_CD4,
+    ICD_DGNS_CD5,
+    ICD_DGNS_VRSN_CD5,
+    ICD_DGNS_CD6,
+    ICD_DGNS_VRSN_CD6,
+    ICD_DGNS_CD7,
+    ICD_DGNS_VRSN_CD7,
+    ICD_DGNS_CD8,
+    ICD_DGNS_VRSN_CD8,
+    ICD_DGNS_CD9,
+    ICD_DGNS_VRSN_CD9,
+    ICD_DGNS_CD10,
+    ICD_DGNS_VRSN_CD10,
+    ICD_DGNS_CD11,
+    ICD_DGNS_VRSN_CD11,
+    ICD_DGNS_CD12,
+    ICD_DGNS_VRSN_CD12,
+    ICD_DGNS_CD13,
+    ICD_DGNS_VRSN_CD13,
+    ICD_DGNS_CD14,
+    ICD_DGNS_VRSN_CD14,
+    ICD_DGNS_CD15,
+    ICD_DGNS_VRSN_CD15,
+    ICD_DGNS_CD16,
+    ICD_DGNS_VRSN_CD16,
+    ICD_DGNS_CD17,
+    ICD_DGNS_VRSN_CD17,
+    ICD_DGNS_CD18,
+    ICD_DGNS_VRSN_CD18,
+    ICD_DGNS_CD19,
+    ICD_DGNS_VRSN_CD19,
+    ICD_DGNS_CD20,
+    ICD_DGNS_VRSN_CD20,
+    ICD_DGNS_CD21,
+    ICD_DGNS_VRSN_CD21,
+    ICD_DGNS_CD22,
+    ICD_DGNS_VRSN_CD22,
+    ICD_DGNS_CD23,
+    ICD_DGNS_VRSN_CD23,
+    ICD_DGNS_CD24,
+    ICD_DGNS_VRSN_CD24,
+    ICD_DGNS_CD25,
+    ICD_DGNS_VRSN_CD25,
+    FST_DGNS_E_CD,
+    FST_DGNS_E_VRSN_CD,
+    ICD_DGNS_E_CD1,
+    ICD_DGNS_E_VRSN_CD1,
+    ICD_DGNS_E_CD2,
+    ICD_DGNS_E_VRSN_CD2,
+    ICD_DGNS_E_CD3,
+    ICD_DGNS_E_VRSN_CD3,
+    ICD_DGNS_E_CD4,
+    ICD_DGNS_E_VRSN_CD4,
+    ICD_DGNS_E_CD5,
+    ICD_DGNS_E_VRSN_CD5,
+    ICD_DGNS_E_CD6,
+    ICD_DGNS_E_VRSN_CD6,
+    ICD_DGNS_E_CD7,
+    ICD_DGNS_E_VRSN_CD7,
+    ICD_DGNS_E_CD8,
+    ICD_DGNS_E_VRSN_CD8,
+    ICD_DGNS_E_CD9,
+    ICD_DGNS_E_VRSN_CD9,
+    ICD_DGNS_E_CD10,
+    ICD_DGNS_E_VRSN_CD10,
+    ICD_DGNS_E_CD11,
+    ICD_DGNS_E_VRSN_CD11,
+    ICD_DGNS_E_CD12,
+    ICD_DGNS_E_VRSN_CD12,
+    CLM_HHA_LUPA_IND_CD,
+    CLM_HHA_RFRL_CD,
+    CLM_HHA_TOT_VISIT_CNT,
+    CLM_ADMSN_DT,
+    FI_DOC_CLM_CNTL_NUM,
+    FI_ORIG_CLM_CNTL_NUM,
+    CLM_LINE_NUM,
+    REV_CNTR,
+    REV_CNTR_DT,
+    REV_CNTR_1ST_ANSI_CD,
+    REV_CNTR_APC_HIPPS_CD,
+    HCPCS_CD,
+    HCPCS_1ST_MDFR_CD,
+    HCPCS_2ND_MDFR_CD,
+    REV_CNTR_PMT_MTHD_IND_CD,
+    REV_CNTR_UNIT_CNT,
+    REV_CNTR_RATE_AMT,
+    REV_CNTR_PMT_AMT_AMT,
+    REV_CNTR_TOT_CHRG_AMT,
+    REV_CNTR_NCVRD_CHRG_AMT,
+    REV_CNTR_DDCTBL_COINSRNC_CD,
+    REV_CNTR_STUS_IND_CD,
+    REV_CNTR_NDC_QTY,
+    REV_CNTR_NDC_QTY_QLFR_CD,
+    RNDRNG_PHYSN_UPIN,
+    RNDRNG_PHYSN_NPI
+  }
+
+  static final HHAFields[][] homeDxFields = {
+    { HHAFields.ICD_DGNS_CD1, HHAFields.ICD_DGNS_VRSN_CD1 },
+    { HHAFields.ICD_DGNS_CD2, HHAFields.ICD_DGNS_VRSN_CD2 },
+    { HHAFields.ICD_DGNS_CD3, HHAFields.ICD_DGNS_VRSN_CD3 },
+    { HHAFields.ICD_DGNS_CD4, HHAFields.ICD_DGNS_VRSN_CD4 },
+    { HHAFields.ICD_DGNS_CD5, HHAFields.ICD_DGNS_VRSN_CD5 },
+    { HHAFields.ICD_DGNS_CD6, HHAFields.ICD_DGNS_VRSN_CD6 },
+    { HHAFields.ICD_DGNS_CD7, HHAFields.ICD_DGNS_VRSN_CD7 },
+    { HHAFields.ICD_DGNS_CD8, HHAFields.ICD_DGNS_VRSN_CD8 },
+    { HHAFields.ICD_DGNS_CD9, HHAFields.ICD_DGNS_VRSN_CD9 },
+    { HHAFields.ICD_DGNS_CD10, HHAFields.ICD_DGNS_VRSN_CD10 },
+    { HHAFields.ICD_DGNS_CD11, HHAFields.ICD_DGNS_VRSN_CD11 },
+    { HHAFields.ICD_DGNS_CD12, HHAFields.ICD_DGNS_VRSN_CD12 },
+    { HHAFields.ICD_DGNS_CD13, HHAFields.ICD_DGNS_VRSN_CD13 },
+    { HHAFields.ICD_DGNS_CD14, HHAFields.ICD_DGNS_VRSN_CD14 },
+    { HHAFields.ICD_DGNS_CD15, HHAFields.ICD_DGNS_VRSN_CD15 },
+    { HHAFields.ICD_DGNS_CD16, HHAFields.ICD_DGNS_VRSN_CD16 },
+    { HHAFields.ICD_DGNS_CD17, HHAFields.ICD_DGNS_VRSN_CD17 },
+    { HHAFields.ICD_DGNS_CD18, HHAFields.ICD_DGNS_VRSN_CD18 },
+    { HHAFields.ICD_DGNS_CD19, HHAFields.ICD_DGNS_VRSN_CD19 },
+    { HHAFields.ICD_DGNS_CD20, HHAFields.ICD_DGNS_VRSN_CD20 },
+    { HHAFields.ICD_DGNS_CD21, HHAFields.ICD_DGNS_VRSN_CD21 },
+    { HHAFields.ICD_DGNS_CD22, HHAFields.ICD_DGNS_VRSN_CD22 },
+    { HHAFields.ICD_DGNS_CD23, HHAFields.ICD_DGNS_VRSN_CD23 },
+    { HHAFields.ICD_DGNS_CD24, HHAFields.ICD_DGNS_VRSN_CD24 },
+    { HHAFields.ICD_DGNS_CD25, HHAFields.ICD_DGNS_VRSN_CD25 }
+  };
+
+  enum HospiceFields {
+    DML_IND,
+    BENE_ID,
+    CLM_ID,
+    CLM_GRP_ID,
+    FINAL_ACTION,
+    NCH_NEAR_LINE_REC_IDENT_CD,
+    NCH_CLM_TYPE_CD,
+    CLM_FROM_DT,
+    CLM_THRU_DT,
+    NCH_WKLY_PROC_DT,
+    FI_CLM_PROC_DT,
+    PRVDR_NUM,
+    CLM_FAC_TYPE_CD,
+    CLM_SRVC_CLSFCTN_TYPE_CD,
+    CLM_FREQ_CD,
+    FI_NUM,
+    CLM_MDCR_NON_PMT_RSN_CD,
+    CLM_PMT_AMT,
+    NCH_PRMRY_PYR_CLM_PD_AMT,
+    NCH_PRMRY_PYR_CD,
+    PRVDR_STATE_CD,
+    ORG_NPI_NUM,
+    AT_PHYSN_UPIN,
+    AT_PHYSN_NPI,
+    PTNT_DSCHRG_STUS_CD,
+    CLM_TOT_CHRG_AMT,
+    NCH_PTNT_STATUS_IND_CD,
+    CLM_UTLZTN_DAY_CNT,
+    NCH_BENE_DSCHRG_DT,
+    PRNCPAL_DGNS_CD,
+    PRNCPAL_DGNS_VRSN_CD,
+    ICD_DGNS_CD1,
+    ICD_DGNS_VRSN_CD1,
+    ICD_DGNS_CD2,
+    ICD_DGNS_VRSN_CD2,
+    ICD_DGNS_CD3,
+    ICD_DGNS_VRSN_CD3,
+    ICD_DGNS_CD4,
+    ICD_DGNS_VRSN_CD4,
+    ICD_DGNS_CD5,
+    ICD_DGNS_VRSN_CD5,
+    ICD_DGNS_CD6,
+    ICD_DGNS_VRSN_CD6,
+    ICD_DGNS_CD7,
+    ICD_DGNS_VRSN_CD7,
+    ICD_DGNS_CD8,
+    ICD_DGNS_VRSN_CD8,
+    ICD_DGNS_CD9,
+    ICD_DGNS_VRSN_CD9,
+    ICD_DGNS_CD10,
+    ICD_DGNS_VRSN_CD10,
+    ICD_DGNS_CD11,
+    ICD_DGNS_VRSN_CD11,
+    ICD_DGNS_CD12,
+    ICD_DGNS_VRSN_CD12,
+    ICD_DGNS_CD13,
+    ICD_DGNS_VRSN_CD13,
+    ICD_DGNS_CD14,
+    ICD_DGNS_VRSN_CD14,
+    ICD_DGNS_CD15,
+    ICD_DGNS_VRSN_CD15,
+    ICD_DGNS_CD16,
+    ICD_DGNS_VRSN_CD16,
+    ICD_DGNS_CD17,
+    ICD_DGNS_VRSN_CD17,
+    ICD_DGNS_CD18,
+    ICD_DGNS_VRSN_CD18,
+    ICD_DGNS_CD19,
+    ICD_DGNS_VRSN_CD19,
+    ICD_DGNS_CD20,
+    ICD_DGNS_VRSN_CD20,
+    ICD_DGNS_CD21,
+    ICD_DGNS_VRSN_CD21,
+    ICD_DGNS_CD22,
+    ICD_DGNS_VRSN_CD22,
+    ICD_DGNS_CD23,
+    ICD_DGNS_VRSN_CD23,
+    ICD_DGNS_CD24,
+    ICD_DGNS_VRSN_CD24,
+    ICD_DGNS_CD25,
+    ICD_DGNS_VRSN_CD25,
+    FST_DGNS_E_CD,
+    FST_DGNS_E_VRSN_CD,
+    ICD_DGNS_E_CD1,
+    ICD_DGNS_E_VRSN_CD1,
+    ICD_DGNS_E_CD2,
+    ICD_DGNS_E_VRSN_CD2,
+    ICD_DGNS_E_CD3,
+    ICD_DGNS_E_VRSN_CD3,
+    ICD_DGNS_E_CD4,
+    ICD_DGNS_E_VRSN_CD4,
+    ICD_DGNS_E_CD5,
+    ICD_DGNS_E_VRSN_CD5,
+    ICD_DGNS_E_CD6,
+    ICD_DGNS_E_VRSN_CD6,
+    ICD_DGNS_E_CD7,
+    ICD_DGNS_E_VRSN_CD7,
+    ICD_DGNS_E_CD8,
+    ICD_DGNS_E_VRSN_CD8,
+    ICD_DGNS_E_CD9,
+    ICD_DGNS_E_VRSN_CD9,
+    ICD_DGNS_E_CD10,
+    ICD_DGNS_E_VRSN_CD10,
+    ICD_DGNS_E_CD11,
+    ICD_DGNS_E_VRSN_CD11,
+    ICD_DGNS_E_CD12,
+    ICD_DGNS_E_VRSN_CD12,
+    CLM_HOSPC_START_DT_ID,
+    BENE_HOSPC_PRD_CNT,
+    FI_DOC_CLM_CNTL_NUM,
+    FI_ORIG_CLM_CNTL_NUM,
+    CLM_LINE_NUM,
+    REV_CNTR,
+    REV_CNTR_DT,
+    HCPCS_CD,
+    HCPCS_1ST_MDFR_CD,
+    HCPCS_2ND_MDFR_CD,
+    REV_CNTR_UNIT_CNT,
+    REV_CNTR_RATE_AMT,
+    REV_CNTR_PRVDR_PMT_AMT,
+    REV_CNTR_BENE_PMT_AMT,
+    REV_CNTR_PMT_AMT_AMT,
+    REV_CNTR_TOT_CHRG_AMT,
+    REV_CNTR_NCVRD_CHRG_AMT,
+    REV_CNTR_DDCTBL_COINSRNC_CD,
+    REV_CNTR_NDC_QTY,
+    REV_CNTR_NDC_QTY_QLFR_CD,
+    RNDRNG_PHYSN_UPIN,
+    RNDRNG_PHYSN_NPI
+  }
+  
+  static final HospiceFields[][] hospiceDxFields = {
+    { HospiceFields.ICD_DGNS_CD1, HospiceFields.ICD_DGNS_VRSN_CD1 },
+    { HospiceFields.ICD_DGNS_CD2, HospiceFields.ICD_DGNS_VRSN_CD2 },
+    { HospiceFields.ICD_DGNS_CD3, HospiceFields.ICD_DGNS_VRSN_CD3 },
+    { HospiceFields.ICD_DGNS_CD4, HospiceFields.ICD_DGNS_VRSN_CD4 },
+    { HospiceFields.ICD_DGNS_CD5, HospiceFields.ICD_DGNS_VRSN_CD5 },
+    { HospiceFields.ICD_DGNS_CD6, HospiceFields.ICD_DGNS_VRSN_CD6 },
+    { HospiceFields.ICD_DGNS_CD7, HospiceFields.ICD_DGNS_VRSN_CD7 },
+    { HospiceFields.ICD_DGNS_CD8, HospiceFields.ICD_DGNS_VRSN_CD8 },
+    { HospiceFields.ICD_DGNS_CD9, HospiceFields.ICD_DGNS_VRSN_CD9 },
+    { HospiceFields.ICD_DGNS_CD10, HospiceFields.ICD_DGNS_VRSN_CD10 },
+    { HospiceFields.ICD_DGNS_CD11, HospiceFields.ICD_DGNS_VRSN_CD11 },
+    { HospiceFields.ICD_DGNS_CD12, HospiceFields.ICD_DGNS_VRSN_CD12 },
+    { HospiceFields.ICD_DGNS_CD13, HospiceFields.ICD_DGNS_VRSN_CD13 },
+    { HospiceFields.ICD_DGNS_CD14, HospiceFields.ICD_DGNS_VRSN_CD14 },
+    { HospiceFields.ICD_DGNS_CD15, HospiceFields.ICD_DGNS_VRSN_CD15 },
+    { HospiceFields.ICD_DGNS_CD16, HospiceFields.ICD_DGNS_VRSN_CD16 },
+    { HospiceFields.ICD_DGNS_CD17, HospiceFields.ICD_DGNS_VRSN_CD17 },
+    { HospiceFields.ICD_DGNS_CD18, HospiceFields.ICD_DGNS_VRSN_CD18 },
+    { HospiceFields.ICD_DGNS_CD19, HospiceFields.ICD_DGNS_VRSN_CD19 },
+    { HospiceFields.ICD_DGNS_CD20, HospiceFields.ICD_DGNS_VRSN_CD20 },
+    { HospiceFields.ICD_DGNS_CD21, HospiceFields.ICD_DGNS_VRSN_CD21 },
+    { HospiceFields.ICD_DGNS_CD22, HospiceFields.ICD_DGNS_VRSN_CD22 },
+    { HospiceFields.ICD_DGNS_CD23, HospiceFields.ICD_DGNS_VRSN_CD23 },
+    { HospiceFields.ICD_DGNS_CD24, HospiceFields.ICD_DGNS_VRSN_CD24 },
+    { HospiceFields.ICD_DGNS_CD25, HospiceFields.ICD_DGNS_VRSN_CD25 }
+  };
+
+  enum SNFFields {
+    DML_IND,
+    BENE_ID,
+    CLM_ID,
+    CLM_GRP_ID,
+    FINAL_ACTION,
+    NCH_NEAR_LINE_REC_IDENT_CD,
+    NCH_CLM_TYPE_CD,
+    CLM_FROM_DT,
+    CLM_THRU_DT,
+    NCH_WKLY_PROC_DT,
+    FI_CLM_PROC_DT,
+    CLAIM_QUERY_CODE,
+    PRVDR_NUM,
+    CLM_FAC_TYPE_CD,
+    CLM_SRVC_CLSFCTN_TYPE_CD,
+    CLM_FREQ_CD,
+    FI_NUM,
+    CLM_MDCR_NON_PMT_RSN_CD,
+    CLM_PMT_AMT,
+    NCH_PRMRY_PYR_CLM_PD_AMT,
+    NCH_PRMRY_PYR_CD,
+    FI_CLM_ACTN_CD,
+    PRVDR_STATE_CD,
+    ORG_NPI_NUM,
+    AT_PHYSN_UPIN,
+    AT_PHYSN_NPI,
+    OP_PHYSN_UPIN,
+    OP_PHYSN_NPI,
+    OT_PHYSN_UPIN,
+    OT_PHYSN_NPI,
+    CLM_MCO_PD_SW,
+    PTNT_DSCHRG_STUS_CD,
+    CLM_PPS_IND_CD,
+    CLM_TOT_CHRG_AMT,
+    CLM_ADMSN_DT,
+    CLM_IP_ADMSN_TYPE_CD,
+    CLM_SRC_IP_ADMSN_CD,
+    NCH_PTNT_STATUS_IND_CD,
+    NCH_BENE_IP_DDCTBL_AMT,
+    NCH_BENE_PTA_COINSRNC_LBLTY_AM,
+    NCH_BENE_BLOOD_DDCTBL_LBLTY_AM,
+    NCH_IP_NCVRD_CHRG_AMT,
+    NCH_IP_TOT_DDCTN_AMT,
+    CLM_PPS_CPTL_FSP_AMT,
+    CLM_PPS_CPTL_OUTLIER_AMT,
+    CLM_PPS_CPTL_DSPRPRTNT_SHR_AMT,
+    CLM_PPS_CPTL_IME_AMT,
+    CLM_PPS_CPTL_EXCPTN_AMT,
+    CLM_PPS_OLD_CPTL_HLD_HRMLS_AMT,
+    CLM_UTLZTN_DAY_CNT,
+    BENE_TOT_COINSRNC_DAYS_CNT,
+    CLM_NON_UTLZTN_DAYS_CNT,
+    NCH_BLOOD_PNTS_FRNSHD_QTY,
+    NCH_QLFYD_STAY_FROM_DT,
+    NCH_QLFYD_STAY_THRU_DT,
+    NCH_VRFD_NCVRD_STAY_FROM_DT,
+    NCH_VRFD_NCVRD_STAY_THRU_DT,
+    NCH_ACTV_OR_CVRD_LVL_CARE_THRU,
+    NCH_BENE_MDCR_BNFTS_EXHTD_DT_I,
+    NCH_BENE_DSCHRG_DT,
+    CLM_DRG_CD,
+    ADMTG_DGNS_CD,
+    ADMTG_DGNS_VRSN_CD,
+    PRNCPAL_DGNS_CD,
+    PRNCPAL_DGNS_VRSN_CD,
+    ICD_DGNS_CD1,
+    ICD_DGNS_VRSN_CD1,
+    ICD_DGNS_CD2,
+    ICD_DGNS_VRSN_CD2,
+    ICD_DGNS_CD3,
+    ICD_DGNS_VRSN_CD3,
+    ICD_DGNS_CD4,
+    ICD_DGNS_VRSN_CD4,
+    ICD_DGNS_CD5,
+    ICD_DGNS_VRSN_CD5,
+    ICD_DGNS_CD6,
+    ICD_DGNS_VRSN_CD6,
+    ICD_DGNS_CD7,
+    ICD_DGNS_VRSN_CD7,
+    ICD_DGNS_CD8,
+    ICD_DGNS_VRSN_CD8,
+    ICD_DGNS_CD9,
+    ICD_DGNS_VRSN_CD9,
+    ICD_DGNS_CD10,
+    ICD_DGNS_VRSN_CD10,
+    ICD_DGNS_CD11,
+    ICD_DGNS_VRSN_CD11,
+    ICD_DGNS_CD12,
+    ICD_DGNS_VRSN_CD12,
+    ICD_DGNS_CD13,
+    ICD_DGNS_VRSN_CD13,
+    ICD_DGNS_CD14,
+    ICD_DGNS_VRSN_CD14,
+    ICD_DGNS_CD15,
+    ICD_DGNS_VRSN_CD15,
+    ICD_DGNS_CD16,
+    ICD_DGNS_VRSN_CD16,
+    ICD_DGNS_CD17,
+    ICD_DGNS_VRSN_CD17,
+    ICD_DGNS_CD18,
+    ICD_DGNS_VRSN_CD18,
+    ICD_DGNS_CD19,
+    ICD_DGNS_VRSN_CD19,
+    ICD_DGNS_CD20,
+    ICD_DGNS_VRSN_CD20,
+    ICD_DGNS_CD21,
+    ICD_DGNS_VRSN_CD21,
+    ICD_DGNS_CD22,
+    ICD_DGNS_VRSN_CD22,
+    ICD_DGNS_CD23,
+    ICD_DGNS_VRSN_CD23,
+    ICD_DGNS_CD24,
+    ICD_DGNS_VRSN_CD24,
+    ICD_DGNS_CD25,
+    ICD_DGNS_VRSN_CD25,
+    FST_DGNS_E_CD,
+    FST_DGNS_E_VRSN_CD,
+    ICD_DGNS_E_CD1,
+    ICD_DGNS_E_VRSN_CD1,
+    ICD_DGNS_E_CD2,
+    ICD_DGNS_E_VRSN_CD2,
+    ICD_DGNS_E_CD3,
+    ICD_DGNS_E_VRSN_CD3,
+    ICD_DGNS_E_CD4,
+    ICD_DGNS_E_VRSN_CD4,
+    ICD_DGNS_E_CD5,
+    ICD_DGNS_E_VRSN_CD5,
+    ICD_DGNS_E_CD6,
+    ICD_DGNS_E_VRSN_CD6,
+    ICD_DGNS_E_CD7,
+    ICD_DGNS_E_VRSN_CD7,
+    ICD_DGNS_E_CD8,
+    ICD_DGNS_E_VRSN_CD8,
+    ICD_DGNS_E_CD9,
+    ICD_DGNS_E_VRSN_CD9,
+    ICD_DGNS_E_CD10,
+    ICD_DGNS_E_VRSN_CD10,
+    ICD_DGNS_E_CD11,
+    ICD_DGNS_E_VRSN_CD11,
+    ICD_DGNS_E_CD12,
+    ICD_DGNS_E_VRSN_CD12,
+    ICD_PRCDR_CD1,
+    ICD_PRCDR_VRSN_CD1,
+    PRCDR_DT1,
+    ICD_PRCDR_CD2,
+    ICD_PRCDR_VRSN_CD2,
+    PRCDR_DT2,
+    ICD_PRCDR_CD3,
+    ICD_PRCDR_VRSN_CD3,
+    PRCDR_DT3,
+    ICD_PRCDR_CD4,
+    ICD_PRCDR_VRSN_CD4,
+    PRCDR_DT4,
+    ICD_PRCDR_CD5,
+    ICD_PRCDR_VRSN_CD5,
+    PRCDR_DT5,
+    ICD_PRCDR_CD6,
+    ICD_PRCDR_VRSN_CD6,
+    PRCDR_DT6,
+    ICD_PRCDR_CD7,
+    ICD_PRCDR_VRSN_CD7,
+    PRCDR_DT7,
+    ICD_PRCDR_CD8,
+    ICD_PRCDR_VRSN_CD8,
+    PRCDR_DT8,
+    ICD_PRCDR_CD9,
+    ICD_PRCDR_VRSN_CD9,
+    PRCDR_DT9,
+    ICD_PRCDR_CD10,
+    ICD_PRCDR_VRSN_CD10,
+    PRCDR_DT10,
+    ICD_PRCDR_CD11,
+    ICD_PRCDR_VRSN_CD11,
+    PRCDR_DT11,
+    ICD_PRCDR_CD12,
+    ICD_PRCDR_VRSN_CD12,
+    PRCDR_DT12,
+    ICD_PRCDR_CD13,
+    ICD_PRCDR_VRSN_CD13,
+    PRCDR_DT13,
+    ICD_PRCDR_CD14,
+    ICD_PRCDR_VRSN_CD14,
+    PRCDR_DT14,
+    ICD_PRCDR_CD15,
+    ICD_PRCDR_VRSN_CD15,
+    PRCDR_DT15,
+    ICD_PRCDR_CD16,
+    ICD_PRCDR_VRSN_CD16,
+    PRCDR_DT16,
+    ICD_PRCDR_CD17,
+    ICD_PRCDR_VRSN_CD17,
+    PRCDR_DT17,
+    ICD_PRCDR_CD18,
+    ICD_PRCDR_VRSN_CD18,
+    PRCDR_DT18,
+    ICD_PRCDR_CD19,
+    ICD_PRCDR_VRSN_CD19,
+    PRCDR_DT19,
+    ICD_PRCDR_CD20,
+    ICD_PRCDR_VRSN_CD20,
+    PRCDR_DT20,
+    ICD_PRCDR_CD21,
+    ICD_PRCDR_VRSN_CD21,
+    PRCDR_DT21,
+    ICD_PRCDR_CD22,
+    ICD_PRCDR_VRSN_CD22,
+    PRCDR_DT22,
+    ICD_PRCDR_CD23,
+    ICD_PRCDR_VRSN_CD23,
+    PRCDR_DT23,
+    ICD_PRCDR_CD24,
+    ICD_PRCDR_VRSN_CD24,
+    PRCDR_DT24,
+    ICD_PRCDR_CD25,
+    ICD_PRCDR_VRSN_CD25,
+    PRCDR_DT25,
+    FI_DOC_CLM_CNTL_NUM,
+    FI_ORIG_CLM_CNTL_NUM,
+    CLM_LINE_NUM,
+    REV_CNTR,
+    HCPCS_CD,
+    REV_CNTR_UNIT_CNT,
+    REV_CNTR_RATE_AMT,
+    REV_CNTR_TOT_CHRG_AMT,
+    REV_CNTR_NCVRD_CHRG_AMT,
+    REV_CNTR_DDCTBL_COINSRNC_CD,
+    REV_CNTR_NDC_QTY,
+    REV_CNTR_NDC_QTY_QLFR_CD,
+    RNDRNG_PHYSN_UPIN,
+    RNDRNG_PHYSN_NPI
+  }
+
+  static final SNFFields[][] snfDxFields = {
+    { SNFFields.ICD_DGNS_CD1, SNFFields.ICD_DGNS_VRSN_CD1 },
+    { SNFFields.ICD_DGNS_CD2, SNFFields.ICD_DGNS_VRSN_CD2 },
+    { SNFFields.ICD_DGNS_CD3, SNFFields.ICD_DGNS_VRSN_CD3 },
+    { SNFFields.ICD_DGNS_CD4, SNFFields.ICD_DGNS_VRSN_CD4 },
+    { SNFFields.ICD_DGNS_CD5, SNFFields.ICD_DGNS_VRSN_CD5 },
+    { SNFFields.ICD_DGNS_CD6, SNFFields.ICD_DGNS_VRSN_CD6 },
+    { SNFFields.ICD_DGNS_CD7, SNFFields.ICD_DGNS_VRSN_CD7 },
+    { SNFFields.ICD_DGNS_CD8, SNFFields.ICD_DGNS_VRSN_CD8 },
+    { SNFFields.ICD_DGNS_CD9, SNFFields.ICD_DGNS_VRSN_CD9 },
+    { SNFFields.ICD_DGNS_CD10, SNFFields.ICD_DGNS_VRSN_CD10 },
+    { SNFFields.ICD_DGNS_CD11, SNFFields.ICD_DGNS_VRSN_CD11 },
+    { SNFFields.ICD_DGNS_CD12, SNFFields.ICD_DGNS_VRSN_CD12 },
+    { SNFFields.ICD_DGNS_CD13, SNFFields.ICD_DGNS_VRSN_CD13 },
+    { SNFFields.ICD_DGNS_CD14, SNFFields.ICD_DGNS_VRSN_CD14 },
+    { SNFFields.ICD_DGNS_CD15, SNFFields.ICD_DGNS_VRSN_CD15 },
+    { SNFFields.ICD_DGNS_CD16, SNFFields.ICD_DGNS_VRSN_CD16 },
+    { SNFFields.ICD_DGNS_CD17, SNFFields.ICD_DGNS_VRSN_CD17 },
+    { SNFFields.ICD_DGNS_CD18, SNFFields.ICD_DGNS_VRSN_CD18 },
+    { SNFFields.ICD_DGNS_CD19, SNFFields.ICD_DGNS_VRSN_CD19 },
+    { SNFFields.ICD_DGNS_CD20, SNFFields.ICD_DGNS_VRSN_CD20 },
+    { SNFFields.ICD_DGNS_CD21, SNFFields.ICD_DGNS_VRSN_CD21 },
+    { SNFFields.ICD_DGNS_CD22, SNFFields.ICD_DGNS_VRSN_CD22 },
+    { SNFFields.ICD_DGNS_CD23, SNFFields.ICD_DGNS_VRSN_CD23 },
+    { SNFFields.ICD_DGNS_CD24, SNFFields.ICD_DGNS_VRSN_CD24 },
+    { SNFFields.ICD_DGNS_CD25, SNFFields.ICD_DGNS_VRSN_CD25 }
+  };
+
+  static final SNFFields[][] snfPxFields = {
+    { SNFFields.ICD_PRCDR_CD1, SNFFields.ICD_PRCDR_VRSN_CD1,  SNFFields.PRCDR_DT1 },
+    { SNFFields.ICD_PRCDR_CD2, SNFFields.ICD_PRCDR_VRSN_CD2, SNFFields.PRCDR_DT2 },
+    { SNFFields.ICD_PRCDR_CD3, SNFFields.ICD_PRCDR_VRSN_CD3, SNFFields.PRCDR_DT3 },
+    { SNFFields.ICD_PRCDR_CD4, SNFFields.ICD_PRCDR_VRSN_CD4, SNFFields.PRCDR_DT4 },
+    { SNFFields.ICD_PRCDR_CD5, SNFFields.ICD_PRCDR_VRSN_CD5, SNFFields.PRCDR_DT5 },
+    { SNFFields.ICD_PRCDR_CD6, SNFFields.ICD_PRCDR_VRSN_CD6, SNFFields.PRCDR_DT6 },
+    { SNFFields.ICD_PRCDR_CD7, SNFFields.ICD_PRCDR_VRSN_CD7, SNFFields.PRCDR_DT7 },
+    { SNFFields.ICD_PRCDR_CD8, SNFFields.ICD_PRCDR_VRSN_CD8, SNFFields.PRCDR_DT8 },
+    { SNFFields.ICD_PRCDR_CD9, SNFFields.ICD_PRCDR_VRSN_CD9, SNFFields.PRCDR_DT9 },
+    { SNFFields.ICD_PRCDR_CD10, SNFFields.ICD_PRCDR_VRSN_CD10, SNFFields.PRCDR_DT10 },
+    { SNFFields.ICD_PRCDR_CD11, SNFFields.ICD_PRCDR_VRSN_CD11, SNFFields.PRCDR_DT11 },
+    { SNFFields.ICD_PRCDR_CD12, SNFFields.ICD_PRCDR_VRSN_CD12, SNFFields.PRCDR_DT12 },
+    { SNFFields.ICD_PRCDR_CD13, SNFFields.ICD_PRCDR_VRSN_CD13, SNFFields.PRCDR_DT13 },
+    { SNFFields.ICD_PRCDR_CD14, SNFFields.ICD_PRCDR_VRSN_CD14, SNFFields.PRCDR_DT14 },
+    { SNFFields.ICD_PRCDR_CD15, SNFFields.ICD_PRCDR_VRSN_CD15, SNFFields.PRCDR_DT15 },
+    { SNFFields.ICD_PRCDR_CD16, SNFFields.ICD_PRCDR_VRSN_CD16, SNFFields.PRCDR_DT16 },
+    { SNFFields.ICD_PRCDR_CD17, SNFFields.ICD_PRCDR_VRSN_CD17, SNFFields.PRCDR_DT17 },
+    { SNFFields.ICD_PRCDR_CD18, SNFFields.ICD_PRCDR_VRSN_CD18, SNFFields.PRCDR_DT18 },
+    { SNFFields.ICD_PRCDR_CD19, SNFFields.ICD_PRCDR_VRSN_CD19, SNFFields.PRCDR_DT19 },
+    { SNFFields.ICD_PRCDR_CD20, SNFFields.ICD_PRCDR_VRSN_CD20, SNFFields.PRCDR_DT20 },
+    { SNFFields.ICD_PRCDR_CD21, SNFFields.ICD_PRCDR_VRSN_CD21, SNFFields.PRCDR_DT21 },
+    { SNFFields.ICD_PRCDR_CD22, SNFFields.ICD_PRCDR_VRSN_CD22, SNFFields.PRCDR_DT22 },
+    { SNFFields.ICD_PRCDR_CD23, SNFFields.ICD_PRCDR_VRSN_CD23, SNFFields.PRCDR_DT23 },
+    { SNFFields.ICD_PRCDR_CD24, SNFFields.ICD_PRCDR_VRSN_CD24, SNFFields.PRCDR_DT24 },
+    { SNFFields.ICD_PRCDR_CD25, SNFFields.ICD_PRCDR_VRSN_CD25, SNFFields.PRCDR_DT25 }
+  };  
+}
